Module rustc_typeck::check::_match

Unstable

Structs

pat_ctxt 	 [Unstable]

Functions

check_dereferencable 	 [Unstable]

check_match 	 [Unstable]

check_pat 	 [Unstable]

check_pat_enum 	 [Unstable]

check_pat_struct 	 [Unstable]

check_struct_pat_fields 	 [Unstable]
path is the AST path item naming the type of this struct. fields is the field patterns of the struct pattern. struct_fields describes the type of each field of the struct. struct_id is the ID of the struct. etc is true if the pattern said '...' and false otherwise.