Module rustc_typeck::check::cast

Code for type-checking cast expressions.

A cast e as U is valid if one of the following holds:

e has type T and T coerces to U; coercion-cast

e has type *T, U is *U_0, and either U_0: Sized or pointer_kind(T) = pointer_kind(U_0); ptr-ptr-cast

e has type *T and U is a numeric type, while T: Sized; ptr-addr-cast

e is an integer and U is *U_0, while U_0: Sized; addr-ptr-cast

e has type T and T and U are any numeric types; numeric-cast

e is a C-like enum and U is an integer type; enum-cast

e has type bool or char and U is an integer; prim-int-cast

e has type u8 and U is char; u8-char-cast

e has type &[T; n] and U is *const T; array-ptr-cast

e is a function pointer type and U has type *T, while T: Sized; fptr-ptr-cast

e is a function pointer type and U is an integer; fptr-addr-cast

where &.T and *T are references of either mutability, and where pointer_kind(T) is the kind of the unsize info in T - the vtable for a trait definition (e.g., fmt::Display or Iterator, not Iterator<Item=u8>) or a length (or () if T: Sized).

Note that lengths are not adjusted when casting raw slices - T: *const [u16] as *const [u8] creates a slice that only includes half of the original memory.

Casting is not transitive, that is, even if e as U1 as U2 is a valid expression, e as U2 is not necessarily so (in fact it will only be valid if U1 coerces to U2).

Structs

CastCheck

Reifies a cast check to be checked once we have full type information for a function context.

Enums

CastCheckResult

CastError

PointerKind 🔒

The kind of pointer and associated metadata (thin, length or vtable) - we only allow casts between fat pointers if their metadata have the same kind.

Functions

check_cast

check_dyn_star_cast 🔒

make_invalid_casting_error 🔒