Function rustc_typeck::check::coercion::try_find_lub

pub fn try_find_lub<'a, 'b, 'tcx, E, I>(fcx: &FnCtxt<'a, 'tcx>, origin: TypeOrigin, exprs: E, prev_ty: Ty<'tcx>, new: &'b Expr) -> RelateResult<'tcx, Ty<'tcx>> where E: Fn() -> I, I: IntoIterator<Item=&'b Expr>

Unstable (rustc_private)

Given some expressions, their known unified type and another expression, tries to unify the types, potentially inserting coercions on any of the provided expressions and returns their LUB (aka "common supertype").