Module rustc_typeck::check::compare_method

Structs

ImplTraitInTraitCollector ğŸ”’

Functions

assoc_item_kind_str ğŸ”’

check_region_bounds_on_impl_item ğŸ”’

check_type_bounds

Validate that ProjectionCandidates created for this associated type will be valid.

collect_trait_impl_trait_tys

compare_const_impl ğŸ”’

compare_generic_param_kinds ğŸ”’

Checks that all parameters in the generics of a given assoc item in a trait impl have the same kind as the respective generic parameter in the trait def.

compare_impl_method ğŸ”’

Checks that a method from an impl conforms to the signature of the same method as declared in the trait.

compare_number_of_generics ğŸ”’

Checks that the number of generics on a given assoc item in a trait impl is the same as the number of generics on the respective assoc item in the trait definition.

compare_number_of_method_arguments ğŸ”’

compare_predicate_entailment ğŸ”’

This function is best explained by example. Consider a trait:

compare_self_type ğŸ”’

compare_synthetic_generics ğŸ”’

compare_ty_impl ğŸ”’

compare_type_predicate_entailment ğŸ”’

The equivalent of compare_predicate_entailment, but for associated types instead of associated functions.

extract_spans_for_error_reporting ğŸ”’