Module rustc_typeck::check::dropck

Structs

SimpleEqRelation 🔒

Functions

check_drop_impl

This function confirms that the Drop implementation identified by drop_impl_did is not any more specialized than the type it is attached to (Issue #8142).

ensure_drop_params_and_item_params_correspond 🔒

ensure_drop_predicates_are_implied_by_item_defn 🔒

Confirms that every predicate imposed by dtor_predicates is implied by assuming the predicates attached to self_type_did.