Enum rustc_typeck::check::Expectation

pub enum Expectation<'tcx> {
    NoExpectation,
    ExpectHasType(Ty<'tcx>),
    ExpectCastableToType(Ty<'tcx>),
    ExpectRvalueLikeUnsized(Ty<'tcx>),
}

Unstable

When type-checking an expression, we propagate downward whatever type hint we are able in the form of an Expectation.

Variants

NoExpectation
Unstable

We know nothing about what type this expression should have.

ExpectHasType
Unstable

This expression should have the type given (or some subtype)

ExpectCastableToType
Unstable

This expression will be cast to the Ty

ExpectRvalueLikeUnsized
Unstable

This rvalue expression will be wrapped in & or Box and coerced to &Ty or Box<Ty>, respectively. Ty is [A] or Trait.