Function rustc_typeck::check::check_instantiable

pub fn check_instantiable(tcx: &ctxt, sp: Span, item_id: NodeId) -> bool

Unstable

Checks whether a type can be created without an instance of itself. This is similar but different from the question of whether a type can be represented. For example, the following type:

enum foo { None, Some(foo) }

is instantiable but is not representable. Similarly, the type

enum foo { Some(@foo) }

is representable, but not instantiable.