Function rustc_typeck::check::compare_impl_method

pub fn compare_impl_method<'tcx>(tcx: &ctxt<'tcx>, impl_m: &Method<'tcx>, impl_m_span: Span, impl_m_body_id: NodeId, trait_m: &Method<'tcx>, impl_trait_ref: &TraitRef<'tcx>)

Unstable

Checks that a method from an impl conforms to the signature of the same method as declared in the trait.

Parameters

impl_m: type of the method we are checking

impl_m_span: span to use for reporting errors

impl_m_body_id: id of the method body

trait_m: the method in the trait

impl_trait_ref: the TraitRef corresponding to the trait implementation