Module rustc_typeck::check::method

Unstable (rustc_private)

Method lookup: the secret sauce of Rust. See README.md.

Reexports

pub use self::MethodError::*;

pub use self::CandidateSource::*;

Structs

NoMatchData 	 [Unstable]

Enums

CandidateSource 	 [Unstable]

MethodError 	 [Unstable]

Functions

exists 	 [Unstable]
Determines whether the type self_ty supports a method name method_name or not.

lookup 	 [Unstable]
Performs method lookup. If lookup is successful, it will return the callee and store an appropriate adjustment for the self-expr. In some cases it may report an error (e.g., invoking the drop method).

lookup_in_trait 	 [Unstable]

lookup_in_trait_adjusted 	 [Unstable]
lookup_in_trait_adjusted is used for overloaded operators. It does a very narrow slice of what the normal probe/confirm path does. In particular, it doesn't really do any probing: it simply constructs an obligation for a particular trait with the given self-type and checks whether that trait is implemented.

report_error 	 [Unstable]

resolve_ufcs 	 [Unstable]

Type Definitions

AllTraitsVec 	 [Unstable]