Function rustc_typeck::constrained_generic_params::parameters_for

pub fn parameters_for<'tcx>(
    t: &impl TypeVisitable<'tcx>,
    include_nonconstraining: bool
) -> Vec<Parameter>

If include_nonconstraining is false, returns the list of parameters that are constrained by t - i.e., the value of each parameter in the list is uniquely determined by t (see RFC 447). If it is true, return the list of parameters whose values are needed in order to constrain ty - these differ, with the latter being a superset, in the presence of projections.