Module rustc_typeck::dep_graph

Unstable (rustc_private)

Modules

debug 	 [Unstable]
Code for debugging the dep-graph.

Structs

DepGraph 	 [Unstable]

DepGraphQuery 	 [Unstable]

DepTask 	 [Unstable]

DepTrackingMap 	 [Unstable]
A DepTrackingMap offers a subset of the Map API and ensures that we make calls to read and write as appropriate. We key the maps with a unique type for brevity.

Enums

DepNode 	 [Unstable]

Traits

DepTrackingMapConfig 	 [Unstable]

Functions

visit_all_items_in_krate 	 [Unstable]
Visit all the items in the krate in some order. When visiting a particular item, first create a dep-node by calling dep_node_fn and push that onto the dep-graph stack of tasks, and also create a read edge from the corresponding AST node. This is used in compiler passes to automatically record the item that they are working on.