Struct rustc_typeck::dep_graph::DepGraph

pub struct DepGraph { /* fields omitted */ }

Unstable (rustc_private)

Methods

impl DepGraph

fn new(enabled: bool) -> DepGraph

Unstable (rustc_private)

fn enabled(&self) -> bool

Unstable (rustc_private)

True if we are actually building a dep-graph. If this returns false, then the other methods on this DepGraph will have no net effect.

fn query(&self) -> DepGraphQuery<DefId>

Unstable (rustc_private)

fn in_ignore(&'graph self) -> IgnoreTask<'graph>

Unstable (rustc_private)

fn in_task(&'graph self, key: DepNode<DefId>) -> DepTask<'graph>

Unstable (rustc_private)

fn with_ignore<OP, R>(&self, op: OP) -> R where OP: FnOnce() -> R

Unstable (rustc_private)

fn with_task<OP, R>(&self, key: DepNode<DefId>, op: OP) -> R where OP: FnOnce() -> R

Unstable (rustc_private)

fn read(&self, v: DepNode<DefId>)

Unstable (rustc_private)

fn write(&self, v: DepNode<DefId>)

Unstable (rustc_private)

fn insert_previous_work_product(&self, v: &Arc<WorkProductId>, data: WorkProduct)

Unstable (rustc_private)

Indicates that a previous work product exists for v. This is invoked during initial start-up based on what nodes are clean (and what files exist in the incr. directory).

fn insert_work_product(&self, v: &Arc<WorkProductId>, data: WorkProduct)

Unstable (rustc_private)

Indicates that we created the given work-product in this run for v. This record will be preserved and loaded in the next run.

fn previous_work_product(&self, v: &Arc<WorkProductId>) -> Option<WorkProduct>

Unstable (rustc_private)

Check whether a previous work product exists for v and, if so, return the path that leads to it. Used to skip doing work.

fn work_products(&self) -> Ref<HashMap<Arc<WorkProductId>, WorkProduct, BuildHasherDefault<FnvHasher>>>

Unstable (rustc_private)

Access the map of work-products created during this run. Only used during saving of the dep-graph.