Struct rustc_typeck::dep_graph::DepTrackingMap

pub struct DepTrackingMap<M> where M: DepTrackingMapConfig { /* fields omitted */ }

Unstable (rustc_private)

A DepTrackingMap offers a subset of the Map API and ensures that we make calls to read and write as appropriate. We key the maps with a unique type for brevity.

Methods

impl<M> DepTrackingMap<M> where M: DepTrackingMapConfig

fn new(graph: DepGraph) -> DepTrackingMap<M>

Unstable (rustc_private)

fn get(&self, k: &M::Key) -> Option<&M::Value>

Unstable (rustc_private)

fn get_mut(&mut self, k: &M::Key) -> Option<&mut M::Value>

Unstable (rustc_private)

fn insert(&mut self, k: M::Key, v: M::Value) -> Option<M::Value>

Unstable (rustc_private)

fn contains_key(&self, k: &M::Key) -> bool

Unstable (rustc_private)

fn keys(&self) -> Vec<M::Key>

Unstable (rustc_private)

fn push<E>(&mut self, k: M::Key, elem: E) where E: Clone, M: DepTrackingMapConfig<Value=Vec<E>>

Unstable (rustc_private)

Append elem to the vector stored for k, creating a new vector if needed. This is considered a write to k.