Enum rustc_typeck::front::map::blocks::Code

pub enum Code<'a> {
    FnLikeCode(FnLikeNode<'a>),
    BlockCode(&'a Block),
}

Unstable (rustc_private)

Carries either an FnLikeNode or a Block, as these are the two constructs that correspond to "code" (as in, something from which we can construct a control-flow graph).

Variants

FnLikeCode
Unstable (rustc_private)

BlockCode
Unstable (rustc_private)

Methods

impl<'a> Code<'a>

fn id(&self) -> u32

Unstable (rustc_private)

fn from_node(node: Node) -> Option<Code>

Unstable (rustc_private)

Attempts to construct a Code from presumed FnLike or Block node input.