Enum rustc_typeck::hir::PatKind

pub enum PatKind {
    Wild,
    Ident(BindingMode, Spanned<Name>, Option<P<Pat>>),
    Struct(Path, P<[Spanned<FieldPat>]>, bool),
    TupleStruct(Path, Option<P<[P<Pat>]>>),
    Path(Path),
    QPath(QSelf, Path),
    Tup(P<[P<Pat>]>),
    Box(P<Pat>),
    Ref(P<Pat>, Mutability),
    Lit(P<Expr>),
    Range(P<Expr>, P<Expr>),
    Vec(P<[P<Pat>]>, Option<P<Pat>>, P<[P<Pat>]>),
}

Unstable (rustc_private)

Variants

Wild
Unstable (rustc_private)

Represents a wildcard pattern (_)

Ident
Unstable (rustc_private)

A PatKind::Ident may either be a new bound variable, or a unit struct/variant pattern, or a const pattern (in the last two cases the third field must be None).

In the unit or const pattern case, the parser can't determine which it is. The resolver determines this, and records this pattern's NodeId in an auxiliary set (of "PatIdents that refer to unit patterns or constants").

Struct
Unstable (rustc_private)

A struct or struct variant pattern, e.g. Variant {x, y, ..}. The bool is true in the presence of a ...

TupleStruct
Unstable (rustc_private)

A tuple struct/variant pattern Variant(x, y, z). "None" means a Variant(..) pattern where we don't bind the fields to names.

Path
Unstable (rustc_private)

A path pattern. Such pattern can be resolved to a unit struct/variant or a constant.

QPath
Unstable (rustc_private)

An associated const named using the qualified path <T>::CONST or <T as Trait>::CONST. Associated consts from inherent impls can be referred to as simply T::CONST, in which case they will end up as PatKind::Path, and the resolver will have to sort that out.

Tup
Unstable (rustc_private)

A tuple pattern (a, b)

Box
Unstable (rustc_private)

A box pattern

Ref
Unstable (rustc_private)

A reference pattern, e.g. &mut (a, b)

Lit
Unstable (rustc_private)

A literal

Range
Unstable (rustc_private)

A range pattern, e.g. 1...2

Vec
Unstable (rustc_private)

[a, b, ..i, y, z] is represented as: PatKind::Vec(box [a, b], Some(i), box [y, z])