Enum rustc_typeck::hir::Ty_

pub enum Ty_ {
    TySlice(P<Ty>),
    TyArray(P<Ty>, BodyId),
    TyPtr(MutTy),
    TyRptr(Option<Lifetime>, MutTy),
    TyBareFn(P<BareFnTy>),
    TyNever,
    TyTup(P<[P<Ty>]>),
    TyPath(QPath),
    TyObjectSum(P<Ty>, P<[TyParamBound]>),
    TyPolyTraitRef(P<[TyParamBound]>),
    TyImplTrait(P<[TyParamBound]>),
    TyTypeof(BodyId),
    TyInfer,
}

🔬 This is a nightly-only experimental API. (rustc_private)

The different kinds of types recognized by the compiler

Variants
TySlice(P<Ty>)

🔬 This is a nightly-only experimental API. (rustc_private)

A variable length slice ([T])

TyArray(P<Ty>, BodyId)

🔬 This is a nightly-only experimental API. (rustc_private)

A fixed length array ([T; n])

TyPtr(MutTy)

🔬 This is a nightly-only experimental API. (rustc_private)

A raw pointer (*const T or *mut T)

TyRptr(Option<Lifetime>, MutTy)

🔬 This is a nightly-only experimental API. (rustc_private)

A reference (&'a T or &'a mut T)

TyBareFn(P<BareFnTy>)

🔬 This is a nightly-only experimental API. (rustc_private)

A bare function (e.g. fn(usize) -> bool)

TyNever

🔬 This is a nightly-only experimental API. (rustc_private)

The never type (!)

TyTup(P<[P<Ty>]>)

🔬 This is a nightly-only experimental API. (rustc_private)

A tuple ((A, B, C, D,...))

TyPath(QPath)

🔬 This is a nightly-only experimental API. (rustc_private)

A path to a type definition (module::module::...::Type), or an associated type, e.g. <Vec<T> as Trait>::Type or <T>::Target.

Type parameters may be stored in each PathSegment.

TyObjectSum(P<Ty>, P<[TyParamBound]>)

🔬 This is a nightly-only experimental API. (rustc_private)

Something like A+B. Note that B must always be a path.

TyPolyTraitRef(P<[TyParamBound]>)

🔬 This is a nightly-only experimental API. (rustc_private)

A type like for<'a> Foo<&'a Bar>

TyImplTrait(P<[TyParamBound]>)

🔬 This is a nightly-only experimental API. (rustc_private)

An impl TraitA+TraitB type.

TyTypeof(BodyId)

🔬 This is a nightly-only experimental API. (rustc_private)

Unused for now

TyInfer

🔬 This is a nightly-only experimental API. (rustc_private)

TyInfer means the type should be inferred instead of it having been specified. This can appear anywhere in a type.