Trait rustc_typeck::hir::intravisit::Visitor

pub trait Visitor<'v> {
    fn nested_visit_map(&mut self) -> Option<&Map<'v>> { ... }
    fn visit_nested_item(&mut self, id: ItemId) { ... }
    fn visit_nested_impl_item(&mut self, id: ImplItemId) { ... }
    fn visit_item(&mut self, i: &'v Item) { ... }
    fn as_deep_visitor(&'s mut self) -> DeepVisitor<'s, Self> { ... }
    fn visit_id(&mut self, _node_id: NodeId) { ... }
    fn visit_name(&mut self, _span: Span, _name: Name) { ... }
    fn visit_mod(&mut self, m: &'v Mod, _s: Span, n: NodeId) { ... }
    fn visit_foreign_item(&mut self, i: &'v ForeignItem) { ... }
    fn visit_local(&mut self, l: &'v Local) { ... }
    fn visit_block(&mut self, b: &'v Block) { ... }
    fn visit_stmt(&mut self, s: &'v Spanned<Stmt_>) { ... }
    fn visit_arm(&mut self, a: &'v Arm) { ... }
    fn visit_pat(&mut self, p: &'v Pat) { ... }
    fn visit_decl(&mut self, d: &'v Spanned<Decl_>) { ... }
    fn visit_expr(&mut self, ex: &'v Expr) { ... }
    fn visit_expr_post(&mut self, _ex: &'v Expr) { ... }
    fn visit_ty(&mut self, t: &'v Ty) { ... }
    fn visit_generics(&mut self, g: &'v Generics) { ... }
    fn visit_where_predicate(&mut self, predicate: &'v WherePredicate) { ... }
    fn visit_fn(&mut self,
                fk: FnKind<'v>,
                fd: &'v FnDecl,
                b: &'v Expr,
                s: Span,
                id: NodeId) { ... }
    fn visit_trait_item(&mut self, ti: &'v TraitItem) { ... }
    fn visit_impl_item(&mut self, ii: &'v ImplItem) { ... }
    fn visit_impl_item_ref(&mut self, ii: &'v ImplItemRef) { ... }
    fn visit_trait_ref(&mut self, t: &'v TraitRef) { ... }
    fn visit_ty_param_bound(&mut self, bounds: &'v TyParamBound) { ... }
    fn visit_poly_trait_ref(&mut self,
                            t: &'v PolyTraitRef,
                            m: &'v TraitBoundModifier) { ... }
    fn visit_variant_data(&mut self,
                          s: &'v VariantData,
                          Name,
                          &'v Generics,
                          _parent_id: NodeId,
                          Span) { ... }
    fn visit_struct_field(&mut self, s: &'v StructField) { ... }
    fn visit_enum_def(&mut self,
                      enum_definition: &'v EnumDef,
                      generics: &'v Generics,
                      item_id: NodeId,
                      Span) { ... }
    fn visit_variant(&mut self,
                     v: &'v Spanned<Variant_>,
                     g: &'v Generics,
                     item_id: NodeId) { ... }
    fn visit_lifetime(&mut self, lifetime: &'v Lifetime) { ... }
    fn visit_lifetime_def(&mut self, lifetime: &'v LifetimeDef) { ... }
    fn visit_path(&mut self, path: &'v Path, _id: NodeId) { ... }
    fn visit_path_list_item(&mut self,
                            prefix: &'v Path,
                            item: &'v Spanned<PathListItem_>) { ... }
    fn visit_path_segment(&mut self,
                          path_span: Span,
                          path_segment: &'v PathSegment) { ... }
    fn visit_path_parameters(&mut self,
                             path_span: Span,
                             path_parameters: &'v PathParameters) { ... }
    fn visit_assoc_type_binding(&mut self, type_binding: &'v TypeBinding) { ... }
    fn visit_attribute(&mut self, _attr: &'v Spanned<Attribute_>) { ... }
    fn visit_macro_def(&mut self, macro_def: &'v MacroDef) { ... }
    fn visit_vis(&mut self, vis: &'v Visibility) { ... }
    fn visit_associated_item_kind(&mut self, kind: &'v AssociatedItemKind) { ... }
    fn visit_defaultness(&mut self, defaultness: &'v Defaultness) { ... }
}

Unstable (rustc_private)

Each method of the Visitor trait is a hook to be potentially overridden. Each method's default implementation recursively visits the substructure of the input via the corresponding walk method; e.g. the visit_mod method by default calls intravisit::walk_mod.

Note that this visitor does NOT visit nested items by default (this is why the module is called intravisit, to distinguish it from the AST's visit module, which acts differently). If you simply want to visit all items in the crate in some order, you should call Crate::visit_all_items. Otherwise, see the comment on visit_nested_item for details on how to visit nested items.

If you want to ensure that your code handles every variant explicitly, you need to override each method. (And you also need to monitor future changes to Visitor in case a new method with a new default implementation gets introduced.)

Provided Methods

fn nested_visit_map(&mut self) -> Option<&Map<'v>>

Unstable (rustc_private)

The default versions of the visit_nested_XXX routines invoke this method to get a map to use; if they get back None, they just skip nested things. Otherwise, they will lookup the nested item-like things in the map and visit it. So the best way to implement a nested visitor is to override this method to return a Map; one advantage of this is that if we add more types of nested things in the future, they will automatically work.

If for some reason you want the nested behavior, but don't have a Map are your disposal: then you should override the visit_nested_XXX methods, and override this method to panic!(). This way, if a new visit_nested_XXX variant is added in the future, we will see the panic in your code and fix it appropriately.

fn visit_nested_item(&mut self, id: ItemId)

Unstable (rustc_private)

Invoked when a nested item is encountered. By default does nothing unless you override nested_visit_map to return Some(_), in which case it will walk the item. You probably don't want to override this method -- instead, override nested_visit_map or use the "shallow" or "deep" visit patterns described on itemlikevisit::ItemLikeVisitor. The only reason to override this method is if you want a nested pattern but cannot supply a Map; see nested_visit_map for advice.

fn visit_nested_impl_item(&mut self, id: ImplItemId)

Unstable (rustc_private)

Like visit_nested_item(), but for impl items. See visit_nested_item() for advice on when to override this method.

fn visit_item(&mut self, i: &'v Item)

Unstable (rustc_private)

Visit the top-level item and (optionally) nested items / impl items. See visit_nested_item for details.

fn as_deep_visitor(&'s mut self) -> DeepVisitor<'s, Self>

Unstable (rustc_private)

When invoking visit_all_item_likes(), you need to supply an item-like visitor. This method converts a "intra-visit" visitor into an item-like visitor that walks the entire tree. If you use this, you probably don't want to process the contents of nested item-like things, since the outer loop will visit them as well.

fn visit_id(&mut self, _node_id: NodeId)

Unstable (rustc_private)

fn visit_name(&mut self, _span: Span, _name: Name)

Unstable (rustc_private)

fn visit_mod(&mut self, m: &'v Mod, _s: Span, n: NodeId)

Unstable (rustc_private)

fn visit_foreign_item(&mut self, i: &'v ForeignItem)

Unstable (rustc_private)

fn visit_local(&mut self, l: &'v Local)

Unstable (rustc_private)

fn visit_block(&mut self, b: &'v Block)

Unstable (rustc_private)

fn visit_stmt(&mut self, s: &'v Spanned<Stmt_>)

Unstable (rustc_private)

fn visit_arm(&mut self, a: &'v Arm)

Unstable (rustc_private)

fn visit_pat(&mut self, p: &'v Pat)

Unstable (rustc_private)

fn visit_decl(&mut self, d: &'v Spanned<Decl_>)

Unstable (rustc_private)

fn visit_expr(&mut self, ex: &'v Expr)

Unstable (rustc_private)

fn visit_expr_post(&mut self, _ex: &'v Expr)

Unstable (rustc_private)

fn visit_ty(&mut self, t: &'v Ty)

Unstable (rustc_private)

fn visit_generics(&mut self, g: &'v Generics)

Unstable (rustc_private)

fn visit_where_predicate(&mut self, predicate: &'v WherePredicate)

Unstable (rustc_private)

fn visit_fn(&mut self,
fk: FnKind<'v>,
fd: &'v FnDecl,
b: &'v Expr,
s: Span,
id: NodeId)

Unstable (rustc_private)

fn visit_trait_item(&mut self, ti: &'v TraitItem)

Unstable (rustc_private)

fn visit_impl_item(&mut self, ii: &'v ImplItem)

Unstable (rustc_private)

fn visit_impl_item_ref(&mut self, ii: &'v ImplItemRef)

Unstable (rustc_private)

fn visit_trait_ref(&mut self, t: &'v TraitRef)

Unstable (rustc_private)

fn visit_ty_param_bound(&mut self, bounds: &'v TyParamBound)

Unstable (rustc_private)

fn visit_poly_trait_ref(&mut self,
t: &'v PolyTraitRef,
m: &'v TraitBoundModifier)

Unstable (rustc_private)

fn visit_variant_data(&mut self,
s: &'v VariantData,
Name,
&'v Generics,
_parent_id: NodeId,
Span)

Unstable (rustc_private)

fn visit_struct_field(&mut self, s: &'v StructField)

Unstable (rustc_private)

fn visit_enum_def(&mut self,
enum_definition: &'v EnumDef,
generics: &'v Generics,
item_id: NodeId,
Span)

Unstable (rustc_private)

fn visit_variant(&mut self,
v: &'v Spanned<Variant_>,
g: &'v Generics,
item_id: NodeId)

Unstable (rustc_private)

fn visit_lifetime(&mut self, lifetime: &'v Lifetime)

Unstable (rustc_private)

fn visit_lifetime_def(&mut self, lifetime: &'v LifetimeDef)

Unstable (rustc_private)

fn visit_path(&mut self, path: &'v Path, _id: NodeId)

Unstable (rustc_private)

fn visit_path_list_item(&mut self,
prefix: &'v Path,
item: &'v Spanned<PathListItem_>)

Unstable (rustc_private)

fn visit_path_segment(&mut self, path_span: Span, path_segment: &'v PathSegment)

Unstable (rustc_private)

fn visit_path_parameters(&mut self,
path_span: Span,
path_parameters: &'v PathParameters)

Unstable (rustc_private)

fn visit_assoc_type_binding(&mut self, type_binding: &'v TypeBinding)

Unstable (rustc_private)

fn visit_attribute(&mut self, _attr: &'v Spanned<Attribute_>)

Unstable (rustc_private)

fn visit_macro_def(&mut self, macro_def: &'v MacroDef)

Unstable (rustc_private)

fn visit_vis(&mut self, vis: &'v Visibility)

Unstable (rustc_private)

fn visit_associated_item_kind(&mut self, kind: &'v AssociatedItemKind)

Unstable (rustc_private)

fn visit_defaultness(&mut self, defaultness: &'v Defaultness)

Unstable (rustc_private)