Struct rustc_typeck::hir::Ident

pub struct Ident {
    pub name: Name,
    pub unhygienic_name: Name,
}

Unstable (rustc_private)

Identifier in HIR

Fields

name
Unstable (rustc_private)

Hygienic name (renamed), should be used by default

unhygienic_name
Unstable (rustc_private)

Unhygienic name (original, not renamed), needed in few places in name resolution

Methods

impl Ident

fn from_name(name: Name) -> Ident

Unstable (rustc_private)

Creates a HIR identifier with both name and unhygienic_name initialized with the argument. Hygiene properties of the created identifier depend entirely on this argument. If the argument is a plain interned string intern("iter"), then the result is unhygienic and can interfere with other entities named "iter". If the argument is a "fresh" name created with gensym("iter"), then the result is hygienic and can't interfere with other entities having the same string as a name.