Struct rustc_typeck::hir::Path

pub struct Path {
    pub span: Span,
    pub global: bool,
    pub segments: P<[PathSegment]>,
}

Unstable (rustc_private)

A "Path" is essentially Rust's notion of a name; for instance: std::cmp::PartialEq . It's represented as a sequence of identifiers, along with a bunch of supporting information.

Fields
span: Span
Unstable (rustc_private)
global: bool
Unstable (rustc_private)

A ::foo path, is relative to the crate root rather than current module (like paths in an import).

segments: P<[PathSegment]>
Unstable (rustc_private)

The segments in the path: the things separated by ::.

Methods

impl Path

fn from_name(s: Span, name: Name) -> Path

Unstable (rustc_private)

Convert a span and an identifier to the corresponding 1-segment path.