Trait rustc_typeck::lint::LateLintPass

pub trait LateLintPass: LintPass {
    fn check_name(&mut self, &LateContext, Span, Name) { ... }
    fn check_crate(&mut self, &LateContext, &Crate) { ... }
    fn check_crate_post(&mut self, &LateContext, &Crate) { ... }
    fn check_mod(&mut self, &LateContext, &Mod, Span, u32) { ... }
    fn check_mod_post(&mut self, &LateContext, &Mod, Span, u32) { ... }
    fn check_foreign_item(&mut self, &LateContext, &ForeignItem) { ... }
    fn check_foreign_item_post(&mut self, &LateContext, &ForeignItem) { ... }
    fn check_item(&mut self, &LateContext, &Item) { ... }
    fn check_item_post(&mut self, &LateContext, &Item) { ... }
    fn check_local(&mut self, &LateContext, &Local) { ... }
    fn check_block(&mut self, &LateContext, &Block) { ... }
    fn check_block_post(&mut self, &LateContext, &Block) { ... }
    fn check_stmt(&mut self, &LateContext, &Spanned<Stmt_>) { ... }
    fn check_arm(&mut self, &LateContext, &Arm) { ... }
    fn check_pat(&mut self, &LateContext, &Pat) { ... }
    fn check_decl(&mut self, &LateContext, &Spanned<Decl_>) { ... }
    fn check_expr(&mut self, &LateContext, &Expr) { ... }
    fn check_expr_post(&mut self, &LateContext, &Expr) { ... }
    fn check_ty(&mut self, &LateContext, &Ty) { ... }
    fn check_generics(&mut self, &LateContext, &Generics) { ... }
    fn check_fn(&mut self, &LateContext, FnKind, &FnDecl, &Block, Span, u32) { ... }
    fn check_fn_post(&mut self, &LateContext, FnKind, &FnDecl, &Block, Span, u32) { ... }
    fn check_trait_item(&mut self, &LateContext, &TraitItem) { ... }
    fn check_trait_item_post(&mut self, &LateContext, &TraitItem) { ... }
    fn check_impl_item(&mut self, &LateContext, &ImplItem) { ... }
    fn check_impl_item_post(&mut self, &LateContext, &ImplItem) { ... }
    fn check_struct_def(&mut self, &LateContext, &VariantData, Name, &Generics, u32) { ... }
    fn check_struct_def_post(&mut self, &LateContext, &VariantData, Name, &Generics, u32) { ... }
    fn check_struct_field(&mut self, &LateContext, &StructField) { ... }
    fn check_variant(&mut self, &LateContext, &Spanned<Variant_>, &Generics) { ... }
    fn check_variant_post(&mut self, &LateContext, &Spanned<Variant_>, &Generics) { ... }
    fn check_lifetime(&mut self, &LateContext, &Lifetime) { ... }
    fn check_lifetime_def(&mut self, &LateContext, &LifetimeDef) { ... }
    fn check_explicit_self(&mut self, &LateContext, &Spanned<SelfKind>) { ... }
    fn check_path(&mut self, &LateContext, &Path, u32) { ... }
    fn check_path_list_item(&mut self, &LateContext, &Spanned<PathListItem_>) { ... }
    fn check_attribute(&mut self, &LateContext, &Spanned<Attribute_>) { ... }
    fn enter_lint_attrs(&mut self, &LateContext, &[Spanned<Attribute_>]) { ... }
    fn exit_lint_attrs(&mut self, &LateContext, &[Spanned<Attribute_>]) { ... }
}

Unstable (rustc_private)

Trait for types providing lint checks.

Each check method checks a single syntax node, and should not invoke methods recursively (unlike Visitor). By default they do nothing.

Provided Methods

fn check_name(&mut self, &LateContext, Span, Name)

Unstable (rustc_private)

fn check_crate(&mut self, &LateContext, &Crate)

Unstable (rustc_private)

fn check_crate_post(&mut self, &LateContext, &Crate)

Unstable (rustc_private)

fn check_mod(&mut self, &LateContext, &Mod, Span, u32)

Unstable (rustc_private)

fn check_mod_post(&mut self, &LateContext, &Mod, Span, u32)

Unstable (rustc_private)

fn check_foreign_item(&mut self, &LateContext, &ForeignItem)

Unstable (rustc_private)

fn check_foreign_item_post(&mut self, &LateContext, &ForeignItem)

Unstable (rustc_private)

fn check_item(&mut self, &LateContext, &Item)

Unstable (rustc_private)

fn check_item_post(&mut self, &LateContext, &Item)

Unstable (rustc_private)

fn check_local(&mut self, &LateContext, &Local)

Unstable (rustc_private)

fn check_block(&mut self, &LateContext, &Block)

Unstable (rustc_private)

fn check_block_post(&mut self, &LateContext, &Block)

Unstable (rustc_private)

fn check_stmt(&mut self, &LateContext, &Spanned<Stmt_>)

Unstable (rustc_private)

fn check_arm(&mut self, &LateContext, &Arm)

Unstable (rustc_private)

fn check_pat(&mut self, &LateContext, &Pat)

Unstable (rustc_private)

fn check_decl(&mut self, &LateContext, &Spanned<Decl_>)

Unstable (rustc_private)

fn check_expr(&mut self, &LateContext, &Expr)

Unstable (rustc_private)

fn check_expr_post(&mut self, &LateContext, &Expr)

Unstable (rustc_private)

fn check_ty(&mut self, &LateContext, &Ty)

Unstable (rustc_private)

fn check_generics(&mut self, &LateContext, &Generics)

Unstable (rustc_private)

fn check_fn(&mut self, &LateContext, FnKind, &FnDecl, &Block, Span, u32)

Unstable (rustc_private)

fn check_fn_post(&mut self, &LateContext, FnKind, &FnDecl, &Block, Span, u32)

Unstable (rustc_private)

fn check_trait_item(&mut self, &LateContext, &TraitItem)

Unstable (rustc_private)

fn check_trait_item_post(&mut self, &LateContext, &TraitItem)

Unstable (rustc_private)

fn check_impl_item(&mut self, &LateContext, &ImplItem)

Unstable (rustc_private)

fn check_impl_item_post(&mut self, &LateContext, &ImplItem)

Unstable (rustc_private)

fn check_struct_def(&mut self, &LateContext, &VariantData, Name, &Generics, u32)

Unstable (rustc_private)

fn check_struct_def_post(&mut self, &LateContext, &VariantData, Name, &Generics, u32)

Unstable (rustc_private)

fn check_struct_field(&mut self, &LateContext, &StructField)

Unstable (rustc_private)

fn check_variant(&mut self, &LateContext, &Spanned<Variant_>, &Generics)

Unstable (rustc_private)

fn check_variant_post(&mut self, &LateContext, &Spanned<Variant_>, &Generics)

Unstable (rustc_private)

fn check_lifetime(&mut self, &LateContext, &Lifetime)

Unstable (rustc_private)

fn check_lifetime_def(&mut self, &LateContext, &LifetimeDef)

Unstable (rustc_private)

fn check_explicit_self(&mut self, &LateContext, &Spanned<SelfKind>)

Unstable (rustc_private)

fn check_path(&mut self, &LateContext, &Path, u32)

Unstable (rustc_private)

fn check_path_list_item(&mut self, &LateContext, &Spanned<PathListItem_>)

Unstable (rustc_private)

fn check_attribute(&mut self, &LateContext, &Spanned<Attribute_>)

Unstable (rustc_private)

fn enter_lint_attrs(&mut self, &LateContext, &[Spanned<Attribute_>])

Unstable (rustc_private)

Called when entering a syntax node that can have lint attributes such as #[allow(...)]. Called with all the attributes of that node.

fn exit_lint_attrs(&mut self, &LateContext, &[Spanned<Attribute_>])

Unstable (rustc_private)

Counterpart to enter_lint_attrs.