Module rustc_typeck::metadata

Unstable (rustc_private)

Modules

common 	 [Unstable]

creader 	 [Unstable]
Validates all used crates and extern libraries and loads their metadata

csearch 	 [Unstable]

cstore 	 [Unstable]

decoder 	 [Unstable]

encoder 	 [Unstable]

filesearch 	 [Unstable]

index 	 [Unstable]

inline 	 [Unstable]

loader 	 [Unstable]
Finds crate binaries and loads their metadata

macro_import 	 [Unstable]
Used by rustc when loading a crate with exported macros.

tydecode 	 [Unstable]

tyencode 	 [Unstable]