Enum rustc_typeck::metadata::inline::InlinedItem

pub enum InlinedItem {
    Item(P<Item>),
    TraitItem(DefId, P<TraitItem>),
    ImplItem(DefId, P<ImplItem>),
    Foreign(P<ForeignItem>),
}

Unstable (rustc_private)

The data we save and restore about an inlined item or method. This is not part of the AST that we parse from a file, but it becomes part of the tree that we trans.

Variants

Item
Unstable (rustc_private)

TraitItem
Unstable (rustc_private)

ImplItem
Unstable (rustc_private)

Foreign
Unstable (rustc_private)

Methods

impl InlinedItem

fn visit<'ast, V>(&'ast self, visitor: &mut V) where V: Visitor<'ast>

Unstable (rustc_private)

fn visit_ids<O>(&self, operation: &mut O) where O: IdVisitingOperation

Unstable (rustc_private)

fn compute_id_range(&self) -> IdRange

Unstable (rustc_private)