Module rustc_typeck::middle::infer

Unstable

See the Book for more information.

Modules

bivariate 	 [Unstable]
Applies the "bivariance relationship" to two types and/or regions. If (A,B) are bivariant then either A <: B or B <: A. It occurs when type/lifetime parameters are unconstrained. Usually this is an error, but we permit it in the specific case where a type parameter is constrained in a where-clause via an associated type.

combine 	 [Unstable]

equate 	 [Unstable]

error_reporting 	 [Unstable]
Error Reporting Code for the inference engine

glb 	 [Unstable]

lattice 	 [Unstable]
Lattice Variables

lub 	 [Unstable]

region_inference 	 [Unstable]
See README.md

resolve 	 [Unstable]

sub 	 [Unstable]

type_variable 	 [Unstable]

unify_key 	 [Unstable]

Structs

CombinedSnapshot 	 [Unstable]

InferCtxt 	 [Unstable]

TypeFreshener 	 [Unstable]

TypeTrace 	 [Unstable]
The trace designates the path through inference that we took to encounter an error or subtyping constraint.

Enums

FixupError 	 [Unstable]

GenericKind 	 [Unstable]

IntVarValue 	 [Unstable]

LateBoundRegionConversionTime 	 [Unstable]
Times when we replace late-bound regions with variables:

ParameterOrigin 	 [Unstable]
Places that type/region parameters can appear.

RegionVariableOrigin 	 [Unstable]
Reasons to create a region inference variable

SubregionOrigin 	 [Unstable]
The origin of a r1 <= r2 constraint.

TypeOrigin 	 [Unstable]
Why did we require that the two types be related?

ValuePairs 	 [Unstable]
See error_reporting.rs for more details

VerifyBound 	 [Unstable]

Functions

can_mk_eqty 	 [Unstable]

can_mk_subty 	 [Unstable]

common_supertype 	 [Unstable]
Computes the least upper-bound of a and b. If this is not possible, reports an error and returns ty::err.

drain_fulfillment_cx 	 [Unstable]
Finishes processes any obligations that remain in the fulfillment context, and then "freshens" and returns result. This is primarily used during normalization and other cases where processing the obligations in fulfill_cx may cause type inference variables that appear in result to be unified, and hence we need to process those obligations to get the complete picture of the type.

drain_fulfillment_cx_or_panic 	 [Unstable]

erase_regions 	 [Unstable]
Returns an equivalent value with all free regions removed (note that late-bound regions remain, because they are important for subtyping, but they are anonymized and normalized as well). This is a stronger, caching version of ty_fold::erase_regions.

fixup_err_to_string 	 [Unstable]

mk_eqty 	 [Unstable]

mk_sub_poly_trait_refs 	 [Unstable]

mk_subr 	 [Unstable]

mk_subty 	 [Unstable]

new_infer_ctxt 	 [Unstable]
errors_will_be_reported is required to proxy to the fulfillment context FIXME -- a better option would be to hold back on modifying the global cache until we know that all dependent obligations are also satisfied. In that case, we could actually remove this boolean flag, and we'd also avoid the problem of squelching duplicate errors that occur across fns.

normalize_associated_type 	 [Unstable]

normalizing_infer_ctxt 	 [Unstable]

Type Definitions

Bound 	 [Unstable]

FixupResult 	 [Unstable]

SkolemizationMap 	 [Unstable]

UnitResult 	 [Unstable]