Enum rustc_typeck::middle::infer::region_inference::RegionResolutionError

pub enum RegionResolutionError<'tcx> {
    ConcreteFailure(SubregionOrigin<'tcx>, Region, Region),
    GenericBoundFailure(SubregionOrigin<'tcx>, GenericKind<'tcx>, Region),
    SubSupConflict(RegionVariableOrigin, SubregionOrigin<'tcx>, Region, SubregionOrigin<'tcx>, Region),
    ProcessedErrors(Vec<RegionVariableOrigin>, Vec<(TypeTrace<'tcx>, TypeError<'tcx>)>, Vec<SameRegions>),
}

Unstable (rustc_private)

Variants

ConcreteFailure
Unstable (rustc_private)

ConcreteFailure(o, a, b):

o requires that a <= b, but this does not hold

GenericBoundFailure
Unstable (rustc_private)

`GenericBoundFailure(p, s, a)

The parameter/associated-type p must be known to outlive the lifetime a (but none of the known bounds are sufficient).

SubSupConflict
Unstable (rustc_private)

SubSupConflict(v, sub_origin, sub_r, sup_origin, sup_r):

Could not infer a value for v because sub_r <= v (due to sub_origin) but v <= sup_r (due to sup_origin) and sub_r <= sup_r does not hold.

ProcessedErrors
Unstable (rustc_private)

For subsets of ConcreteFailure and SubSupConflict, we can derive more specific errors message by suggesting to the user where they should put a lifetime. In those cases we process and put those errors into ProcessedErrors before we do any reporting.