Struct rustc_typeck::middle::infer::InferCtxt

pub struct InferCtxt<'a, 'tcx> where 'tcx: 'a {
    pub tcx: &'a TyCtxt<'tcx>,
    pub tables: &'a RefCell<Tables<'tcx>>,
    pub parameter_environment: ParameterEnvironment<'a, 'tcx>,
    pub reported_trait_errors: RefCell<HashSet<TraitErrorKey<'tcx>, BuildHasherDefault<FnvHasher>>>,
    // some fields omitted
}

Unstable (rustc_private)

Fields

tcx
Unstable (rustc_private)

tables
Unstable (rustc_private)

parameter_environment
Unstable (rustc_private)

reported_trait_errors
Unstable (rustc_private)

Methods

impl<'a, 'tcx> InferCtxt<'a, 'tcx>

fn freshen<T>(&self, t: T) -> T where T: TypeFoldable<'tcx>

Unstable (rustc_private)

fn type_var_diverges(&'a self, ty: &TyS) -> bool

Unstable (rustc_private)

fn freshener(&'b self) -> TypeFreshener<'b, 'tcx>

Unstable (rustc_private)

fn type_is_unconstrained_numeric(&'a self, ty: &TyS) -> UnconstrainedNumeric

Unstable (rustc_private)

fn default(&self, ty: &'tcx TyS<'tcx>) -> Option<Default<'tcx>>

Unstable (rustc_private)

Returns a type variable's default fallback if any exists. A default must be attached to the variable when created, if it is created without a default, this will return None.

This code does not apply to integral or floating point variables, only to use declared defaults.

See new_ty_var_with_default to create a type variable with a default. See type_variable::Default for details about what a default entails.

fn unsolved_variables(&self) -> Vec<&'tcx TyS<'tcx>>

Unstable (rustc_private)

fn equate(&'a self, a_is_expected: bool, trace: TypeTrace<'tcx>) -> Equate<'a, 'tcx>

Unstable (rustc_private)

fn sub(&'a self, a_is_expected: bool, trace: TypeTrace<'tcx>) -> Sub<'a, 'tcx>

Unstable (rustc_private)

fn lub(&'a self, a_is_expected: bool, trace: TypeTrace<'tcx>) -> Lub<'a, 'tcx>

Unstable (rustc_private)

fn glb(&'a self, a_is_expected: bool, trace: TypeTrace<'tcx>) -> Glb<'a, 'tcx>

Unstable (rustc_private)

fn commit_unconditionally<R, F>(&self, f: F) -> R where F: FnOnce() -> R

Unstable (rustc_private)

Execute f and commit the bindings

fn commit_if_ok<T, E, F>(&self, f: F) -> Result<T, E> where F: FnOnce(&CombinedSnapshot) -> Result<T, E>

Unstable (rustc_private)

Execute f and commit the bindings if closure f returns Ok(_)

fn commit_regions_if_ok<T, E, F>(&self, f: F) -> Result<T, E> where F: FnOnce() -> Result<T, E>

Unstable (rustc_private)

Execute f and commit only the region bindings if successful. The function f must be very careful not to leak any non-region variables that get created.

fn probe<R, F>(&self, f: F) -> R where F: FnOnce(&CombinedSnapshot) -> R

Unstable (rustc_private)

Execute f then unroll any bindings it creates

fn add_given(&self, sub: FreeRegion, sup: RegionVid)

Unstable (rustc_private)

fn sub_types(&self, a_is_expected: bool, origin: TypeOrigin, a: &'tcx TyS<'tcx>, b: &'tcx TyS<'tcx>) -> Result<(), TypeError<'tcx>>

Unstable (rustc_private)

fn eq_types(&self, a_is_expected: bool, origin: TypeOrigin, a: &'tcx TyS<'tcx>, b: &'tcx TyS<'tcx>) -> Result<(), TypeError<'tcx>>

Unstable (rustc_private)

fn eq_trait_refs(&self, a_is_expected: bool, origin: TypeOrigin, a: TraitRef<'tcx>, b: TraitRef<'tcx>) -> Result<(), TypeError<'tcx>>

Unstable (rustc_private)

fn sub_poly_trait_refs(&self, a_is_expected: bool, origin: TypeOrigin, a: Binder<TraitRef<'tcx>>, b: Binder<TraitRef<'tcx>>) -> Result<(), TypeError<'tcx>>

Unstable (rustc_private)

fn skolemize_late_bound_regions<T>(&self, value: &Binder<T>, snapshot: &CombinedSnapshot) -> (T, HashMap<BoundRegion, Region, BuildHasherDefault<FnvHasher>>) where T: TypeFoldable<'tcx>

Unstable (rustc_private)

See higher_ranked::skolemize_late_bound_regions

fn leak_check(&self, skol_map: &HashMap<BoundRegion, Region, BuildHasherDefault<FnvHasher>>, snapshot: &CombinedSnapshot) -> Result<(), TypeError<'tcx>>

Unstable (rustc_private)

See higher_ranked::leak_check

fn plug_leaks<T>(&self, skol_map: HashMap<BoundRegion, Region, BuildHasherDefault<FnvHasher>>, snapshot: &CombinedSnapshot, value: &T) -> T where T: TypeFoldable<'tcx>

Unstable (rustc_private)

See higher_ranked::plug_leaks

fn equality_predicate(&self, span: Span, predicate: &Binder<EquatePredicate<'tcx>>) -> Result<(), TypeError<'tcx>>

Unstable (rustc_private)

fn region_outlives_predicate(&self, span: Span, predicate: &Binder<OutlivesPredicate<Region, Region>>) -> Result<(), TypeError<'tcx>>

Unstable (rustc_private)

fn next_ty_var_id(&self, diverging: bool) -> TyVid

Unstable (rustc_private)

fn next_ty_var(&self) -> &'tcx TyS<'tcx>

Unstable (rustc_private)

fn next_ty_var_with_default(&self, default: Option<Default<'tcx>>) -> &'tcx TyS<'tcx>

Unstable (rustc_private)

fn next_diverging_ty_var(&self) -> &'tcx TyS<'tcx>

Unstable (rustc_private)

fn next_ty_vars(&self, n: usize) -> Vec<&'tcx TyS<'tcx>>

Unstable (rustc_private)

fn next_int_var_id(&self) -> IntVid

Unstable (rustc_private)

fn next_float_var_id(&self) -> FloatVid

Unstable (rustc_private)

fn next_region_var(&self, origin: RegionVariableOrigin) -> Region

Unstable (rustc_private)

fn region_vars_for_defs(&self, span: Span, defs: &[RegionParameterDef]) -> Vec<Region>

Unstable (rustc_private)

fn type_vars_for_defs(&self, span: Span, space: ParamSpace, substs: &mut Substs<'tcx>, defs: &[TypeParameterDef<'tcx>])

Unstable (rustc_private)

fn fresh_substs_for_generics(&self, span: Span, generics: &Generics<'tcx>) -> Substs<'tcx>

Unstable (rustc_private)

Given a set of generics defined on a type or impl, returns a substitution mapping each type/region parameter to a fresh inference variable.

fn fresh_substs_for_trait(&self, span: Span, generics: &Generics<'tcx>, self_ty: &'tcx TyS<'tcx>) -> Substs<'tcx>

Unstable (rustc_private)

Given a set of generics defined on a trait, returns a substitution mapping each output type/region parameter to a fresh inference variable, and mapping the self type to self_ty.

fn fresh_bound_region(&self, debruijn: DebruijnIndex) -> Region

Unstable (rustc_private)

fn adjust_expr_ty(&self, expr: &Expr, adjustment: Option<&AutoAdjustment<'tcx>>) -> &'tcx TyS<'tcx>

Unstable (rustc_private)

Apply adjustment to the type of expr

fn errors_since_creation(&self) -> bool

Unstable (rustc_private)

fn node_type(&self, id: u32) -> &'tcx TyS<'tcx>

Unstable (rustc_private)

fn expr_ty(&self, ex: &Expr) -> &'tcx TyS<'tcx>

Unstable (rustc_private)

fn resolve_regions_and_report_errors(&self, free_regions: &FreeRegionMap, subject_node_id: u32)

Unstable (rustc_private)

fn ty_to_string(&self, t: &'tcx TyS<'tcx>) -> String

Unstable (rustc_private)

fn tys_to_string(&self, ts: &[&'tcx TyS<'tcx>]) -> String

Unstable (rustc_private)

fn trait_ref_to_string(&self, t: &TraitRef<'tcx>) -> String

Unstable (rustc_private)

fn shallow_resolve(&self, typ: &'tcx TyS<'tcx>) -> &'tcx TyS<'tcx>

Unstable (rustc_private)

fn resolve_type_vars_if_possible<T>(&self, value: &T) -> T where T: TypeFoldable<'tcx>

Unstable (rustc_private)

Where possible, replaces type/int/float variables in value with their final value. Note that region variables are unaffected. If a type variable has not been unified, it is left as is. This is an idempotent operation that does not affect inference state in any way and so you can do it at will.

fn resolve_type_and_region_vars_if_possible<T>(&self, value: &T) -> T where T: TypeFoldable<'tcx>

Unstable (rustc_private)

fn fully_resolve<T>(&self, value: &T) -> Result<T, FixupError> where T: TypeFoldable<'tcx>

Unstable (rustc_private)

Attempts to resolve all type/region variables in value. Region inference must have been run already (e.g., by calling resolve_regions_and_report_errors). If some variable was never unified, an Err results.

This method is idempotent, but it not typically not invoked except during the writeback phase.

fn type_error_message_str<M>(&self, sp: Span, mk_msg: M, actual_ty: String, err: Option<&TypeError<'tcx>>) where M: FnOnce(Option<String>, String) -> String

Unstable (rustc_private)

fn type_error_struct_str<M>(&self, sp: Span, mk_msg: M, actual_ty: String, err: Option<&TypeError<'tcx>>) -> DiagnosticBuilder<'tcx> where M: FnOnce(Option<String>, String) -> String

Unstable (rustc_private)

fn type_error_message_str_with_expected<M>(&self, sp: Span, mk_msg: M, expected_ty: Option<&'tcx TyS<'tcx>>, actual_ty: String, err: Option<&TypeError<'tcx>>) where M: FnOnce(Option<String>, String) -> String

Unstable (rustc_private)

fn type_error_struct_str_with_expected<M>(&self, sp: Span, mk_msg: M, expected_ty: Option<&'tcx TyS<'tcx>>, actual_ty: String, err: Option<&TypeError<'tcx>>) -> DiagnosticBuilder<'tcx> where M: FnOnce(Option<String>, String) -> String

Unstable (rustc_private)

fn type_error_message<M>(&self, sp: Span, mk_msg: M, actual_ty: &'tcx TyS<'tcx>, err: Option<&TypeError<'tcx>>) where M: FnOnce(String) -> String

Unstable (rustc_private)

fn type_error_struct<M>(&self, sp: Span, mk_msg: M, actual_ty: &'tcx TyS<'tcx>, err: Option<&TypeError<'tcx>>) -> DiagnosticBuilder<'tcx> where M: FnOnce(String) -> String

Unstable (rustc_private)

fn report_mismatched_types(&self, origin: TypeOrigin, expected: &'tcx TyS<'tcx>, actual: &'tcx TyS<'tcx>, err: TypeError<'tcx>)

Unstable (rustc_private)

fn report_conflicting_default_types(&self, span: Span, expected: Default<'tcx>, actual: Default<'tcx>)

Unstable (rustc_private)

fn replace_late_bound_regions_with_fresh_var<T>(&self, span: Span, lbrct: LateBoundRegionConversionTime, value: &Binder<T>) -> (T, HashMap<BoundRegion, Region, BuildHasherDefault<FnvHasher>>) where T: TypeFoldable<'tcx>

Unstable (rustc_private)

fn verify_generic_bound(&self, origin: SubregionOrigin<'tcx>, kind: GenericKind<'tcx>, a: Region, bound: VerifyBound)

Unstable (rustc_private)

See verify_generic_bound method in region_inference

fn can_equate<'b, T>(&'b self, a: &T, b: &T) -> Result<(), TypeError<'tcx>> where T: Relate<'b, 'tcx> + Debug

Unstable (rustc_private)

fn node_ty(&self, id: u32) -> Result<&'tcx TyS<'tcx>, ()>

Unstable (rustc_private)

fn expr_ty_adjusted(&self, expr: &Expr) -> Result<&'tcx TyS<'tcx>, ()>

Unstable (rustc_private)

fn tables_are_tcx_tables(&self) -> bool

Unstable (rustc_private)

fn type_moves_by_default(&self, ty: &'tcx TyS<'tcx>, span: Span) -> bool

Unstable (rustc_private)

fn node_method_ty(&self, method_call: MethodCall) -> Option<&'tcx TyS<'tcx>>

Unstable (rustc_private)

fn node_method_id(&self, method_call: MethodCall) -> Option<DefId>

Unstable (rustc_private)

fn adjustments(&self) -> Ref<HashMap<u32, AutoAdjustment<'tcx>, BuildHasherDefault<FnvHasher>>>

Unstable (rustc_private)

fn is_method_call(&self, id: u32) -> bool

Unstable (rustc_private)

fn temporary_scope(&self, rvalue_id: u32) -> Option<CodeExtent>

Unstable (rustc_private)

fn upvar_capture(&self, upvar_id: UpvarId) -> Option<UpvarCapture>

Unstable (rustc_private)

fn param_env(&'b self) -> &'b ParameterEnvironment<'b, 'tcx>

Unstable (rustc_private)

fn closure_kind(&self, def_id: DefId) -> Option<ClosureKind>

Unstable (rustc_private)

fn closure_type(&self, def_id: DefId, substs: &ClosureSubsts<'tcx>) -> ClosureTy<'tcx>

Unstable (rustc_private)