Module rustc_typeck::middle::region

Unstable (rustc_private)

This file actually contains two passes related to regions. The first pass builds up the scope_map, which describes the parent links in the region hierarchy. The second pass infers which types must be region parameterized.

Most of the documentation on regions can be found in middle/infer/region_inference/README.md

Structs

BlockRemainder 	 [Unstable]
Represents a subscope of block for a binding that is introduced by block.stmts[first_statement_index]. Such subscopes represent a suffix of the block. Note that each subscope does not include the initializer expression, if any, for the statement indexed by first_statement_index.

CallSiteScopeData 	 [Unstable]
extent of call-site for a function/method.

CodeExtent 	 [Unstable]

Context 	 [Unstable]

RegionMaps 	 [Unstable]
The region maps encode information about region relationships.

Enums

CodeExtentData 	 [Unstable]
CodeExtent represents a statically-describable extent that can be used to bound the lifetime/region for values.

Constants

DUMMY_CODE_EXTENT 	 [Unstable]
A placeholder used in trans to stand for real code extents

ROOT_CODE_EXTENT 	 [Unstable]
The root of everything. I should be using NonZero or profiling instead of this (probably).

Functions

resolve_crate 	 [Unstable]

resolve_inlined_item 	 [Unstable]