Struct rustc_typeck::middle::stability::Index

pub struct Index<'tcx> {
    // some fields omitted
}

Unstable (rustc_private)

A stability index, giving the stability level for items and methods.

Methods

impl<'tcx> Index<'tcx>

fn build(&mut self, tcx: &ctxt<'tcx>, krate: &Crate, export_map: &HashSet<u32, DefaultState<FnvHasher>>)

Unstable (rustc_private)

Construct the stability index for a crate being compiled.

fn new(krate: &Crate) -> Index

Unstable (rustc_private)