Module rustc_typeck::middle::substUnstable

Structs

EnumeratedItems

SeparateVecsPerParamSpace
The split function converts one VecPerParamSpace into this SeparateVecsPerParamSpace structure.

Substs
A substitution mapping type/region parameters to new values. We identify each in-scope parameter by an index and a parameter space (which indices where the parameter is defined; see ParamSpace).

VecPerParamSpace
Vector of things sorted by param space. Used to keep the set of things declared on the type, self, or method distinct.

Enums

ParamSpace

RegionSubsts
Represents the values to use when substituting lifetime parameters. If the value is ErasedRegions, then this subst is occurring during trans, and all region parameters will be replaced with ty::ReStatic.

Traits

Subst