Struct rustc_typeck::middle::subst::Substs

pub struct Substs<'tcx> {
    pub types: VecPerParamSpace<&'tcx TyS<'tcx>>,
    pub regions: RegionSubsts,
}

Unstable

A substitution mapping type/region parameters to new values. We identify each in-scope parameter by an index and a parameter space (which indices where the parameter is defined; see ParamSpace).

Fields

types
Unstable

regions
Unstable

Methods

impl<'tcx> Substs<'tcx>

fn new(t: VecPerParamSpace<&'tcx TyS<'tcx>>, r: VecPerParamSpace<Region>) -> Substs<'tcx>

Unstable

fn new_type(t: Vec<&'tcx TyS<'tcx>>, r: Vec<Region>) -> Substs<'tcx>

Unstable

fn new_trait(t: Vec<&'tcx TyS<'tcx>>, r: Vec<Region>, s: &'tcx TyS<'tcx>) -> Substs<'tcx>

Unstable

fn erased(t: VecPerParamSpace<&'tcx TyS<'tcx>>) -> Substs<'tcx>

Unstable

fn empty() -> Substs<'tcx>

Unstable

fn trans_empty() -> Substs<'tcx>

Unstable

fn is_noop(&self) -> bool

Unstable

fn type_for_def(&self, ty_param_def: &TypeParameterDef) -> &'tcx TyS<'tcx>

Unstable

fn self_ty(&self) -> Option<&'tcx TyS<'tcx>>

Unstable

fn with_self_ty(&self, self_ty: &'tcx TyS<'tcx>) -> Substs<'tcx>

Unstable

fn erase_regions(self) -> Substs<'tcx>

Unstable

fn regions(&'a self) -> &'a VecPerParamSpace<Region>

Unstable

Since ErasedRegions are only to be used in trans, most of the compiler can use this method to easily access the set of region substitutions.

fn mut_regions(&'a mut self) -> &'a mut VecPerParamSpace<Region>

Unstable

Since ErasedRegions are only to be used in trans, most of the compiler can use this method to easily access the set of region substitutions.

fn with_method(self, m_types: Vec<&'tcx TyS<'tcx>>, m_regions: Vec<Region>) -> Substs<'tcx>

Unstable