Module rustc_typeck::middle::traits

Unstable (rustc_private)

Trait Resolution. See the Book for more.

Structs

DerivedObligationCause 	 [Unstable]

FulfillmentError 	 [Unstable]

Obligation 	 [Unstable]
An Obligation represents some trait reference (e.g. int:Eq) for which the vtable must be found. The process of finding a vtable is called "resolving" the Obligation. This process consists of either identifying an impl (e.g., impl Eq for int) that provides the required vtable, or else finding a bound that is in scope. The eventual result is usually a Selection (defined below).

ObligationCause 	 [Unstable]
Why did we incur this obligation? Used for error reporting.

RFC1214Warning 	 [Unstable]
This marker is used in some caches to record whether the predicate, if it is found to be false, will yield a warning (due to RFC1214) or an error. We separate these two cases in the cache so that if we see the same predicate twice, first resulting in a warning, and next resulting in an error, we still report the error, rather than considering it a duplicate.

VtableBuiltinData 	 [Unstable]

VtableClosureData 	 [Unstable]

VtableDefaultImplData 	 [Unstable]

VtableImplData 	 [Unstable]
Identifies a particular impl in the source, along with a set of substitutions from the impl's type/lifetime parameters. The nested vector corresponds to the nested obligations attached to the impl's type parameters.

VtableObjectData 	 [Unstable]
A vtable for some object-safe trait Foo automatically derived for the object type Foo.

Enums

FulfillmentErrorCode 	 [Unstable]

ObjectSafetyViolation 	 [Unstable]

ObligationCauseCode 	 [Unstable]

SelectionError 	 [Unstable]

Vtable 	 [Unstable]
Given the successful resolution of an obligation, the Vtable indicates where the vtable comes from. Note that while we call this a "vtable", it does not necessarily indicate dynamic dispatch at runtime. Vtable instances just tell the compiler where to find methods, but in generic code those methods are typically statically dispatched -- only when an object is constructed is a Vtable instance reified into an actual vtable.

Functions

fully_normalize 	 [Unstable]

normalize_param_env_or_error 	 [Unstable]
Normalizes the parameter environment, reporting errors if they occur.

predicates_for_generics 	 [Unstable]
Creates predicate obligations from the generic bounds.

type_known_to_meet_builtin_bound 	 [Unstable]
Determines whether the type ty is known to meet bound and returns true if so. Returns false if ty either does not meet bound or is not known to meet bound (note that this is conservative towards no impl, which is the opposite of the evaluate methods).

Type Definitions

Obligations 	 [Unstable]

PredicateObligation 	 [Unstable]

PredicateObligations 	 [Unstable]

Selection 	 [Unstable]

SelectionResult 	 [Unstable]

TraitObligation 	 [Unstable]

TraitObligations 	 [Unstable]