Function rustc_typeck::middle::traits::is_vtable_safe_method

pub fn is_vtable_safe_method(tcx: &ctxt<'tcx>, trait_def_id: DefId, method: &Method<'tcx>) -> bool

Unstable

We say a method is vtable safe if it can be invoked on a trait object. Note that object-safe traits can have some non-vtable-safe methods, so long as they require Self:Sized or otherwise ensure that they cannot be used when Self=Trait.