Struct rustc_typeck::middle::traits::specialization_graph::Graph

pub struct Graph {
    // some fields omitted
}

Unstable (rustc_private)

A per-trait graph of impls in specialization order. At the moment, this graph forms a tree rooted with the trait itself, with all other nodes representing impls, and parent-child relationships representing specializations.

The graph provides two key services:

Construction, which implicitly checks for overlapping impls (i.e., impls that overlap but where neither specializes the other -- an artifact of the simple "chain" rule.

Parent extraction. In particular, the graph can give you the immediate parents of a given specializing impl, which is needed for extracting default items amongst other thigns. In the simple "chain" rule, every impl has at most one parent.

Methods

impl Graph

fn new() -> Graph

Unstable (rustc_private)

fn insert(&mut self, tcx: &'a TyCtxt<'tcx>, impl_def_id: DefId) -> Result<(), Overlap<'a, 'tcx>>

Unstable (rustc_private)

Insert a local impl into the specialization graph. If an existing impl conflicts with it (has overlap, but neither specializes the other), information about the area of overlap is returned in the Err.

fn record_impl_from_cstore(&mut self, parent: DefId, child: DefId)

Unstable (rustc_private)

Insert cached metadata mapping from a child impl back to its parent.

fn parent(&self, child: DefId) -> DefId

Unstable (rustc_private)

The parent of a given impl, which is the def id of the trait when the impl is a "specialization root".