Struct rustc_typeck::middle::traits::Obligation

pub struct Obligation<'tcx, T> {
    pub cause: ObligationCause<'tcx>,
    pub recursion_depth: usize,
    pub predicate: T,
}

Unstable (rustc_private)

An Obligation represents some trait reference (e.g. int:Eq) for which the vtable must be found. The process of finding a vtable is called "resolving" the Obligation. This process consists of either identifying an impl (e.g., impl Eq for int) that provides the required vtable, or else finding a bound that is in scope. The eventual result is usually a Selection (defined below).

Fields

cause
Unstable (rustc_private)

recursion_depth
Unstable (rustc_private)

predicate
Unstable (rustc_private)

Methods

impl<'tcx, O> Obligation<'tcx, O>

fn new(cause: ObligationCause<'tcx>, trait_ref: O) -> Obligation<'tcx, O>

Unstable (rustc_private)

fn misc(span: Span, body_id: u32, trait_ref: O) -> Obligation<'tcx, O>

Unstable (rustc_private)

fn with<P>(&self, value: P) -> Obligation<'tcx, P>

Unstable (rustc_private)