Struct rustc_typeck::middle::traits::VtableImplData

pub struct VtableImplData<'tcx, N> {
    pub impl_def_id: DefId,
    pub substs: &'tcx Substs<'tcx>,
    pub nested: Vec<N>,
}

Unstable (rustc_private)

Identifies a particular impl in the source, along with a set of substitutions from the impl's type/lifetime parameters. The nested vector corresponds to the nested obligations attached to the impl's type parameters.

The type parameter N indicates the type used for "nested obligations" that are required by the impl. During type check, this is Obligation, as one might expect. During trans, however, this is (), because trans only requires a shallow resolution of an impl, and nested obligations are satisfied later.

Fields

impl_def_id
Unstable (rustc_private)

substs
Unstable (rustc_private)

nested
Unstable (rustc_private)