Struct rustc_typeck::middle::traits::VtableObjectData

pub struct VtableObjectData<'tcx> {
    pub upcast_trait_ref: Binder<TraitRef<'tcx>>,
    pub vtable_base: usize,
}

Unstable (rustc_private)

A vtable for some object-safe trait Foo automatically derived for the object type Foo.

Fields

upcast_trait_ref
Unstable (rustc_private)

Foo upcast to the obligation trait. This will be some supertrait of Foo.

vtable_base
Unstable (rustc_private)

The vtable is formed by concatenating together the method lists of the base object trait and all supertraits; this is the start of upcast_trait_ref's methods in that vtable.