Enum rustc_typeck::middle::ty::Predicate

pub enum Predicate<'tcx> {
    Trait(Binder<TraitPredicate<'tcx>>),
    Equate(Binder<EquatePredicate<'tcx>>),
    RegionOutlives(Binder<OutlivesPredicate<Region, Region>>),
    TypeOutlives(Binder<OutlivesPredicate<&'tcx TyS<'tcx>, Region>>),
    Projection(Binder<ProjectionPredicate<'tcx>>),
    WellFormed(&'tcx TyS<'tcx>),
    ObjectSafe(DefId),
}

Unstable (rustc_private)

Variants

Trait
Unstable (rustc_private)

Corresponds to where Foo : Bar<A,B,C>. Foo here would be the Self type of the trait reference and A, B, and C would be the parameters in the TypeSpace.

Equate
Unstable (rustc_private)

where T1 == T2.

RegionOutlives
Unstable (rustc_private)

where 'a : 'b

TypeOutlives
Unstable (rustc_private)

where T : 'a

Projection
Unstable (rustc_private)

where ::Name == X, approximately. See ProjectionPredicate struct for details.

WellFormed
Unstable (rustc_private)

no syntax: T WF

ObjectSafe
Unstable (rustc_private)

trait must be object-safe

Methods

impl<'tcx> Predicate<'tcx>

fn subst_supertrait(&self, tcx: &ctxt<'tcx>, trait_ref: &Binder<TraitRef<'tcx>>) -> Predicate<'tcx>

Unstable (rustc_private)

Performs a substitution suitable for going from a poly-trait-ref to supertraits that must hold if that poly-trait-ref holds. This is slightly different from a normal substitution in terms of what happens with bound regions. See lengthy comment below for details.

impl<'tcx> Predicate<'tcx>

fn walk_tys(&self) -> IntoIter<&'tcx TyS<'tcx>>

Unstable (rustc_private)

Iterates over the types in this predicate. Note that in all cases this is skipping over a binder, so late-bound regions with depth 0 are bound by the predicate.

fn to_opt_poly_trait_ref(&self) -> Option<Binder<TraitRef<'tcx>>>

Unstable (rustc_private)