Enum rustc_typeck::middle::ty::Region

pub enum Region {
    ReEarlyBound(EarlyBoundRegion),
    ReLateBound(DebruijnIndex, BoundRegion),
    ReFree(FreeRegion),
    ReScope(CodeExtent),
    ReStatic,
    ReInfer(InferRegion),
    ReEmpty,
}

Unstable

Representation of regions:

Variants

ReEarlyBound
Unstable

ReLateBound
Unstable

ReFree
Unstable

When checking a function body, the types of all arguments and so forth that refer to bound region parameters are modified to refer to free region parameters.

ReScope
Unstable

A concrete region naming some statically determined extent (e.g. an expression or sequence of statements) within the current function.

ReStatic
Unstable

Static data that has an "infinite" lifetime. Top in the region lattice.

ReInfer
Unstable

A region variable. Should not exist after typeck.

ReEmpty
Unstable

Empty lifetime is for data that is never accessed. Bottom in the region lattice. We treat ReEmpty somewhat specially; at least right now, we do not generate instances of it during the GLB computations, but rather generate an error instead. This is to improve error messages. The only way to get an instance of ReEmpty is to have a region variable with no constraints.

Methods

impl Region

fn is_global(&self) -> bool

Unstable

fn is_bound(&self) -> bool

Unstable

fn escapes_depth(&self, depth: u32) -> bool

Unstable