Enum rustc_typeck::middle::ty::sty

pub enum sty<'tcx> {
    ty_bool,
    ty_char,
    ty_int(IntTy),
    ty_uint(UintTy),
    ty_float(FloatTy),
    ty_enum(DefId, &'tcx Substs<'tcx>),
    ty_uniq(&'tcx TyS<'tcx>),
    ty_str,
    ty_vec(&'tcx TyS<'tcx>, Option<usize>),
    ty_ptr(mt<'tcx>),
    ty_rptr(&'tcx Region, mt<'tcx>),
    ty_bare_fn(Option<DefId>, &'tcx BareFnTy<'tcx>),
    ty_trait(Box<TyTrait<'tcx>>),
    ty_struct(DefId, &'tcx Substs<'tcx>),
    ty_closure(DefId, &'tcx Substs<'tcx>),
    ty_tup(Vec<&'tcx TyS<'tcx>>),
    ty_projection(ProjectionTy<'tcx>),
    ty_param(ParamTy),
    ty_infer(InferTy),
    ty_err,
}

Unstable

Variants

ty_bool
Unstable

ty_char
Unstable

ty_int
Unstable

ty_uint
Unstable

ty_float
Unstable

ty_enum
Unstable

Substs here, possibly against intuition, may contain ty_params. That is, even after substitution it is possible that there are type variables. This happens when the ty_enum corresponds to an enum definition and not a concrete use of it. To get the correct ty_enum from the tcx, use the NodeId from the ast::Ty and look it up in the ast_ty_to_ty_cache. This is probably true for ty_struct as well.

ty_uniq
Unstable

ty_str
Unstable

ty_vec
Unstable

ty_ptr
Unstable

ty_rptr
Unstable

ty_bare_fn
Unstable

ty_trait
Unstable

ty_struct
Unstable

ty_closure
Unstable

ty_tup
Unstable

ty_projection
Unstable

ty_param
Unstable

ty_infer
Unstable

ty_err
Unstable