Function rustc_typeck::middle::ty::hash_crate_independent

pub fn hash_crate_independent(tcx: &ctxt<'tcx>, ty: &'tcx TyS<'tcx>, svh: &Svh) -> u64

Unstable

Creates a hash of the type Ty which will be the same no matter what crate context it's calculated within. This is used by the type_id intrinsic.