Function rustc_typeck::middle::ty::type_err_to_str

pub fn type_err_to_str(cx: &ctxt<'tcx>, err: &type_err<'tcx>) -> String

Unstable

Explains the source of a type err in a short, human readable way. This is meant to be placed in parentheses after some larger message. You should also invoke note_and_explain_type_err() afterwards to present additional details, particularly when it comes to lifetime-related errors.