Struct rustc_typeck::middle::ty::Tables

pub struct Tables<'tcx> {
    pub node_types: HashMap<u32, &'tcx TyS<'tcx>, DefaultState<FnvHasher>>,
    pub item_substs: HashMap<u32, ItemSubsts<'tcx>, DefaultState<FnvHasher>>,
    pub adjustments: HashMap<u32, AutoAdjustment<'tcx>, DefaultState<FnvHasher>>,
    pub method_map: HashMap<MethodCall, MethodCallee<'tcx>, DefaultState<FnvHasher>>,
    pub upvar_capture_map: HashMap<UpvarId, UpvarCapture, DefaultState<FnvHasher>>,
    pub closure_tys: HashMap<DefId, ClosureTy<'tcx>, DefaultState<FnvHasher>>,
    pub closure_kinds: HashMap<DefId, ClosureKind, DefaultState<FnvHasher>>,
}

Unstable

Fields

node_types
Unstable

Stores the types for various nodes in the AST. Note that this table is not guaranteed to be populated until after typeck. See typeck::check::fn_ctxt for details.

item_substs
Unstable

Stores the type parameters which were substituted to obtain the type of this node. This only applies to nodes that refer to entities parameterized by type parameters, such as generic fns, types, or other items.

adjustments
Unstable

method_map
Unstable

upvar_capture_map
Unstable

Borrows

closure_tys
Unstable

Records the type of each closure. The def ID is the ID of the expression defining the closure.

closure_kinds
Unstable

Records the type of each closure. The def ID is the ID of the expression defining the closure.

Methods

impl<'tcx> Tables<'tcx>

fn empty() -> Tables<'tcx>

Unstable