Struct rustc_typeck::middle::ty::ctxt

pub struct ctxt<'tcx> {
    pub dep_graph: DepGraph,
    pub types: CommonTypes<'tcx>,
    pub sess: &'tcx Session,
    pub def_map: RefCell<HashMap<u32, PathResolution, BuildHasherDefault<FnvHasher>>>,
    pub named_region_map: HashMap<u32, DefRegion, BuildHasherDefault<FnvHasher>>,
    pub region_maps: RegionMaps,
    pub tables: RefCell<Tables<'tcx>>,
    pub impl_or_trait_items: RefCell<DepTrackingMap<ImplOrTraitItems<'tcx>>>,
    pub trait_item_def_ids: RefCell<DepTrackingMap<TraitItemDefIds<'tcx>>>,
    pub impl_trait_refs: RefCell<DepTrackingMap<ImplTraitRefs<'tcx>>>,
    pub trait_defs: RefCell<DepTrackingMap<TraitDefs<'tcx>>>,
    pub adt_defs: RefCell<DepTrackingMap<AdtDefs<'tcx>>>,
    pub predicates: RefCell<DepTrackingMap<Predicates<'tcx>>>,
    pub super_predicates: RefCell<DepTrackingMap<Predicates<'tcx>>>,
    pub map: Map<'tcx>,
    pub freevars: RefCell<HashMap<u32, Vec<Freevar>, BuildHasherDefault<FnvHasher>>>,
    pub tcache: RefCell<DepTrackingMap<Tcache<'tcx>>>,
    pub rcache: RefCell<HashMap<CReaderCacheKey, &'tcx TyS<'tcx>, BuildHasherDefault<FnvHasher>>>,
    pub tc_cache: RefCell<HashMap<&'tcx TyS<'tcx>, TypeContents, BuildHasherDefault<FnvHasher>>>,
    pub ast_ty_to_ty_cache: RefCell<HashMap<u32, &'tcx TyS<'tcx>, BuildHasherDefault<FnvHasher>>>,
    pub ty_param_defs: RefCell<HashMap<u32, TypeParameterDef<'tcx>, BuildHasherDefault<FnvHasher>>>,
    pub normalized_cache: RefCell<HashMap<&'tcx TyS<'tcx>, &'tcx TyS<'tcx>, BuildHasherDefault<FnvHasher>>>,
    pub lang_items: LanguageItems,
    pub item_variance_map: RefCell<DepTrackingMap<ItemVariances<'tcx>>>,
    pub variance_computed: Cell<bool>,
    pub inherent_impls: RefCell<DepTrackingMap<InherentImpls<'tcx>>>,
    pub impl_items: RefCell<DepTrackingMap<ImplItems<'tcx>>>,
    pub used_unsafe: RefCell<HashSet<u32, BuildHasherDefault<FnvHasher>>>,
    pub used_mut_nodes: RefCell<HashSet<u32, BuildHasherDefault<FnvHasher>>>,
    pub populated_external_types: RefCell<HashSet<DefId, BuildHasherDefault<FnvHasher>>>,
    pub populated_external_primitive_impls: RefCell<HashSet<DefId, BuildHasherDefault<FnvHasher>>>,
    pub extern_const_statics: RefCell<HashMap<DefId, u32, BuildHasherDefault<FnvHasher>>>,
    pub extern_const_fns: RefCell<HashMap<DefId, u32, BuildHasherDefault<FnvHasher>>>,
    pub node_lint_levels: RefCell<HashMap<(u32, LintId), (Level, LintSource), BuildHasherDefault<FnvHasher>>>,
    pub transmute_restrictions: RefCell<Vec<TransmuteRestriction<'tcx>>>,
    pub stability: RefCell<Index<'tcx>>,
    pub selection_cache: SelectionCache<'tcx>,
    pub evaluation_cache: EvaluationCache<'tcx>,
    pub fulfilled_predicates: RefCell<GlobalFulfilledPredicates<'tcx>>,
    pub const_qualif_map: RefCell<HashMap<u32, ConstQualif, BuildHasherDefault<FnvHasher>>>,
    pub custom_coerce_unsized_kinds: RefCell<HashMap<DefId, CustomCoerceUnsized, BuildHasherDefault<FnvHasher>>>,
    pub cast_kinds: RefCell<HashMap<u32, CastKind, BuildHasherDefault<FnvHasher>>>,
    pub fragment_infos: RefCell<HashMap<DefId, Vec<FragmentInfo>, BuildHasherDefault<FnvHasher>>>,
    // some fields omitted
}

Unstable (rustc_private)

The data structure to keep track of all the information that typechecker generates so that so that it can be reused and doesn't have to be redone later on.

Fields

dep_graph
Unstable (rustc_private)

types
Unstable (rustc_private)

Common types, pre-interned for your convenience.

sess
Unstable (rustc_private)

def_map
Unstable (rustc_private)

named_region_map
Unstable (rustc_private)

region_maps
Unstable (rustc_private)

tables
Unstable (rustc_private)

impl_or_trait_items
Unstable (rustc_private)

Maps from a trait item to the trait item "descriptor"

trait_item_def_ids
Unstable (rustc_private)

Maps from a trait def-id to a list of the def-ids of its trait items

impl_trait_refs
Unstable (rustc_private)

trait_defs
Unstable (rustc_private)

adt_defs
Unstable (rustc_private)

predicates
Unstable (rustc_private)

Maps from the def-id of an item (trait/struct/enum/fn) to its associated predicates.

super_predicates
Unstable (rustc_private)

Maps from the def-id of a trait to the list of super-predicates. This is a subset of the full list of predicates. We store these in a separate map because we must evaluate them even during type conversion, often before the full predicates are available (note that supertraits have additional acyclicity requirements).

map
Unstable (rustc_private)

freevars
Unstable (rustc_private)

tcache
Unstable (rustc_private)

rcache
Unstable (rustc_private)

tc_cache
Unstable (rustc_private)

ast_ty_to_ty_cache
Unstable (rustc_private)

ty_param_defs
Unstable (rustc_private)

normalized_cache
Unstable (rustc_private)

lang_items
Unstable (rustc_private)

item_variance_map
Unstable (rustc_private)

Maps from def-id of a type or region parameter to its (inferred) variance.

variance_computed
Unstable (rustc_private)

True if the variance has been computed yet; false otherwise.

inherent_impls
Unstable (rustc_private)

Maps a DefId of a type to a list of its inherent impls. Contains implementations of methods that are inherent to a type. Methods in these implementations don't need to be exported.

impl_items
Unstable (rustc_private)

Maps a DefId of an impl to a list of its items. Note that this contains all of the impls that we know about, including ones in other crates. It's not clear that this is the best way to do it.

used_unsafe
Unstable (rustc_private)

Set of used unsafe nodes (functions or blocks). Unsafe nodes not present in this set can be warned about.

used_mut_nodes
Unstable (rustc_private)

Set of nodes which mark locals as mutable which end up getting used at some point. Local variable definitions not in this set can be warned about.

populated_external_types
Unstable (rustc_private)

The set of external nominal types whose implementations have been read. This is used for lazy resolution of methods.

populated_external_primitive_impls
Unstable (rustc_private)

The set of external primitive types whose implementations have been read. FIXME(arielb1): why is this separate from populated_external_types?

extern_const_statics
Unstable (rustc_private)

These caches are used by const_eval when decoding external constants.

extern_const_fns
Unstable (rustc_private)

node_lint_levels
Unstable (rustc_private)

transmute_restrictions
Unstable (rustc_private)

The types that must be asserted to be the same size for transmute to be valid. We gather up these restrictions in the intrinsicck pass and check them in trans.

stability
Unstable (rustc_private)

Maps any item's def-id to its stability index.

selection_cache
Unstable (rustc_private)

Caches the results of trait selection. This cache is used for things that do not have to do with the parameters in scope.

evaluation_cache
Unstable (rustc_private)

Caches the results of trait evaluation. This cache is used for things that do not have to do with the parameters in scope. Merge this with selection_cache?

fulfilled_predicates
Unstable (rustc_private)

A set of predicates that have been fulfilled somewhere. This is used to avoid duplicate work. Predicates are only added to this set when they mention only "global" names (i.e., no type or lifetime parameters).

const_qualif_map
Unstable (rustc_private)

Maps Expr NodeId's to their constant qualification.

custom_coerce_unsized_kinds
Unstable (rustc_private)

Caches CoerceUnsized kinds for impls on custom types.

cast_kinds
Unstable (rustc_private)

Maps a cast expression to its kind. This is keyed on the from expression of the cast, not the cast itself.

fragment_infos
Unstable (rustc_private)

Maps Fn items to a collection of fragment infos.

The main goal is to identify data (each of which may be moved or assigned) whose subparts are not moved nor assigned (i.e. their state is unfragmented) and corresponding ast nodes where the path to that data is moved or assigned.

In the long term, unfragmented values will have their destructor entirely driven by a single stack-local drop-flag, and their parents, the collections of the unfragmented values (or more simply, "fragmented values"), are mapped to the corresponding collections of stack-local drop-flags.

(However, in the short term that is not the case; e.g. some unfragmented paths still need to be zeroed, namely when they reference parent data from an outer scope that was not entirely moved, and therefore that needs to be zeroed so that we do not get double-drop when we hit the end of the parent scope.)

Also: currently the table solely holds keys for node-ids of unfragmented values (see FragmentInfo enum definition), but longer-term we will need to also store mappings from fragmented data to the set of unfragmented pieces that constitute it.

Methods

impl<'tcx> ctxt<'tcx>

fn note_and_explain_region(&self, err: &mut DiagnosticBuilder, prefix: &str, region: Region, suffix: &str)

Unstable (rustc_private)

impl<'tcx> ctxt<'tcx>

fn note_and_explain_type_err(&self, db: &mut DiagnosticBuilder, err: &TypeError<'tcx>, sp: Span)

Unstable (rustc_private)

impl<'tcx> ctxt<'tcx>

fn collect_regions<T>(&self, value: &T, region_set: &mut HashSet<Region, BuildHasherDefault<FnvHasher>>) -> bool where T: TypeFoldable<'tcx>

Unstable (rustc_private)

Collects the free and escaping regions in value into region_set. Returns whether any late-bound regions were skipped

fn fold_regions<T, F>(&self, value: &T, skipped_regions: &mut bool, f: F) -> T where T: TypeFoldable<'tcx>, F: FnMut(Region, u32) -> Region

Unstable (rustc_private)

Folds the escaping and free regions in value using f, and sets skipped_regions to true if any late-bound region was found and skipped.

impl<'tcx> ctxt<'tcx>

fn replace_late_bound_regions<T, F>(&self, value: &Binder<T>, f: F) -> (T, HashMap<BoundRegion, Region, BuildHasherDefault<FnvHasher>>) where T: TypeFoldable<'tcx>, F: FnMut(BoundRegion) -> Region

Unstable (rustc_private)

fn liberate_late_bound_regions<T>(&self, all_outlive_scope: CodeExtent, value: &Binder<T>) -> T where T: TypeFoldable<'tcx>

Unstable (rustc_private)

Replace any late-bound regions bound in value with free variants attached to scope-id scope_id.

fn flatten_late_bound_regions<T>(&self, bound2_value: &Binder<Binder<T>>) -> Binder<T> where T: TypeFoldable<'tcx>

Unstable (rustc_private)

Flattens two binding levels into one. So for<'a> for<'b> Foo becomes for<'a,'b> Foo.

fn no_late_bound_regions<T>(&self, value: &Binder<T>) -> Option<T> where T: TypeFoldable<'tcx>

Unstable (rustc_private)

fn erase_late_bound_regions<T>(&self, value: &Binder<T>) -> T where T: TypeFoldable<'tcx>

Unstable (rustc_private)

Replace any late-bound regions bound in value with 'static. Useful in trans but also method lookup and a few other places where precise region relationships are not required.

fn anonymize_late_bound_regions<T>(&self, sig: &Binder<T>) -> Binder<T> where T: TypeFoldable<'tcx>

Unstable (rustc_private)

Rewrite any late-bound regions so that they are anonymous. Region numbers are assigned starting at 1 and increasing monotonically in the order traversed by the fold operation.

The chief purpose of this function is to canonicalize regions so that two FnSigs or TraitRefs which are equivalent up to region naming will become structurally identical. For example, for<'a, 'b> fn(&'a isize, &'b isize) and for<'a, 'b> fn(&'b isize, &'a isize) will become identical after anonymization.

impl<'tcx> ctxt<'tcx>

fn erase_regions<T>(&self, value: &T) -> T where T: TypeFoldable<'tcx>

Unstable (rustc_private)

Returns an equivalent value with all free regions removed (note that late-bound regions remain, because they are important for subtyping, but they are anonymized and normalized as well)..

impl<'tcx> ctxt<'tcx>

fn pat_contains_ref_binding(&self, pat: &Pat) -> Option<Mutability>

Unstable (rustc_private)

fn arm_contains_ref_binding(&self, arm: &Arm) -> Option<Mutability>

Unstable (rustc_private)

fn positional_element_ty(&self, ty: &'tcx TyS<'tcx>, i: usize, variant: Option<DefId>) -> Option<&'tcx TyS<'tcx>>

Unstable (rustc_private)

Returns the type of element at index i in tuple or tuple-like type t. For an enum t, variant is None only if t is a univariant enum.

fn named_element_ty(&self, ty: &'tcx TyS<'tcx>, n: Name, variant: Option<DefId>) -> Option<&'tcx TyS<'tcx>>

Unstable (rustc_private)

Returns the type of element at field n in struct or struct-like type t. For an enum t, variant must be some def id.

fn enum_repr_type(&self, opt_hint: Option<&ReprAttr>) -> (IntType, &'tcx TyS<'tcx>)

Unstable (rustc_private)

Returns (normalized_type, ty), where normalized_type is the IntType representation of one of {i64,i32,i16,i8,u64,u32,u16,u8}, and ty is the original type (i.e. may include isize or usize).

fn struct_tail(&self, ty: &'tcx TyS<'tcx>) -> &'tcx TyS<'tcx>

Unstable (rustc_private)

Returns the deeply last field of nested structures, or the same type, if not a structure at all. Corresponds to the only possible unsized field, and its type can be used to determine unsizing strategy.

fn struct_lockstep_tails(&self, source: &'tcx TyS<'tcx>, target: &'tcx TyS<'tcx>) -> (&'tcx TyS<'tcx>, &'tcx TyS<'tcx>)

Unstable (rustc_private)

Same as applying struct_tail on source and target, but only keeps going as long as the two types are instances of the same structure definitions. For (Foo<Foo<T>>, Foo<Trait>), the result will be (Foo<T>, Trait), whereas struct_tail produces T, and Trait, respectively.

fn eval_repeat_count(&self, count_expr: &Expr) -> usize

Unstable (rustc_private)

Returns the repeat count for a repeating vector expression.

fn required_region_bounds(&self, erased_self_ty: &'tcx TyS<'tcx>, predicates: Vec<Predicate<'tcx>>) -> Vec<Region>

Unstable (rustc_private)

Given a set of predicates that apply to an object type, returns the region bounds that the (erased) Self type must outlive. Precisely because the Self type is erased, the parameter erased_self_ty must be supplied to indicate what type has been used to represent Self in the predicates themselves. This should really be a unique type; FreshTy(0) is a popular choice.

NB: in some cases, particularly around higher-ranked bounds, this function returns a kind of conservative approximation. That is, all regions returned by this function are definitely required, but there may be other region bounds that are not returned, as well as requirements like for<'a> T: 'a.

Requires that trait definitions have been processed so that we can elaborate predicates and walk supertraits.

fn hash_crate_independent(&self, ty: &'tcx TyS<'tcx>, svh: &Svh) -> u64

Unstable (rustc_private)

Creates a hash of the type Ty which will be the same no matter what crate context it's calculated within. This is used by the type_id intrinsic.

fn is_adt_dtorck(&self, adt: &'tcx AdtDefData<'tcx, 'static>) -> bool

Unstable (rustc_private)

Returns true if this ADT is a dtorck type.

Invoking the destructor of a dtorck type during usual cleanup (e.g. the glue emitted for stack unwinding) requires all lifetimes in the type-structure of adt to strictly outlive the adt value itself.

If adt is not dtorck, then the adt's destructor can be invoked even when there are lifetimes in the type-structure of adt that do not strictly outlive the adt value itself. (This allows programs to make cyclic structures without resorting to unasfe means; see RFCs 769 and 1238).

impl<'tcx> ctxt<'tcx>

fn get_impl_method(&self, impl_def_id: DefId, substs: Substs<'tcx>, name: Name) -> ImplMethod<'tcx>

Unstable (rustc_private)

impl<'tcx> ctxt<'tcx>

fn type_parameter_def(&self, node_id: u32) -> TypeParameterDef<'tcx>

Unstable (rustc_private)

fn node_types(&self) -> Ref<HashMap<u32, &'tcx TyS<'tcx>, BuildHasherDefault<FnvHasher>>>

Unstable (rustc_private)

fn node_type_insert(&self, id: u32, ty: &'tcx TyS<'tcx>)

Unstable (rustc_private)

fn intern_trait_def(&self, def: TraitDef<'tcx>) -> &'tcx TraitDef<'tcx>

Unstable (rustc_private)

fn alloc_trait_def(&self, def: TraitDef<'tcx>) -> &'tcx TraitDef<'tcx>

Unstable (rustc_private)

fn intern_adt_def(&self, did: DefId, kind: AdtKind, variants: Vec<VariantDefData<'tcx, 'tcx>>) -> &'tcx AdtDefData<'tcx, 'tcx>

Unstable (rustc_private)

fn intern_stability(&self, stab: Stability) -> &'tcx Stability

Unstable (rustc_private)

fn store_free_region_map(&self, id: u32, map: FreeRegionMap)

Unstable (rustc_private)

fn free_region_map(&self, id: u32) -> FreeRegionMap

Unstable (rustc_private)

fn lift<T>(&self, value: &T) -> Option<T::Lifted> where T: Lift<'tcx> + ?Sized

Unstable (rustc_private)

fn create_and_enter<F, R>(s: &'tcx Session, arenas: &'tcx CtxtArenas<'tcx>, def_map: RefCell<HashMap<u32, PathResolution, BuildHasherDefault<FnvHasher>>>, named_region_map: HashMap<u32, DefRegion, BuildHasherDefault<FnvHasher>>, map: Map<'tcx>, freevars: HashMap<u32, Vec<Freevar>, BuildHasherDefault<FnvHasher>>, region_maps: RegionMaps, lang_items: LanguageItems, stability: Index<'tcx>, f: F) -> R where F: FnOnce(&ctxt<'tcx>) -> R

Unstable (rustc_private)

Create a type context and call the closure with a &ty::ctxt reference to the context. The closure enforces that the type context and any interned value (types, substs, etc.) can only be used while ty::tls has a valid reference to the context, to allow formatting values that need it.

impl<'tcx> ctxt<'tcx>

fn print_debug_stats(&self)

Unstable (rustc_private)

impl<'tcx> ctxt<'tcx>

fn mk_substs(&self, substs: Substs<'tcx>) -> &'tcx Substs<'tcx>

Unstable (rustc_private)

fn safe_to_unsafe_fn_ty(&self, bare_fn: &BareFnTy<'tcx>) -> &'tcx TyS<'tcx>

Unstable (rustc_private)

Create an unsafe fn ty based on a safe fn ty.

fn mk_bare_fn(&self, bare_fn: BareFnTy<'tcx>) -> &'tcx BareFnTy<'tcx>

Unstable (rustc_private)

fn mk_region(&self, region: Region) -> &'tcx Region

Unstable (rustc_private)

fn mk_ty(&self, st: TypeVariants<'tcx>) -> &'tcx TyS<'tcx>

Unstable (rustc_private)

fn mk_mach_int(&self, tm: IntTy) -> &'tcx TyS<'tcx>

Unstable (rustc_private)

fn mk_mach_uint(&self, tm: UintTy) -> &'tcx TyS<'tcx>

Unstable (rustc_private)

fn mk_mach_float(&self, tm: FloatTy) -> &'tcx TyS<'tcx>

Unstable (rustc_private)

fn mk_str(&self) -> &'tcx TyS<'tcx>

Unstable (rustc_private)

fn mk_static_str(&self) -> &'tcx TyS<'tcx>

Unstable (rustc_private)

fn mk_enum(&self, def: &'tcx AdtDefData<'tcx, 'static>, substs: &'tcx Substs<'tcx>) -> &'tcx TyS<'tcx>

Unstable (rustc_private)

fn mk_box(&self, ty: &'tcx TyS<'tcx>) -> &'tcx TyS<'tcx>

Unstable (rustc_private)

fn mk_ptr(&self, tm: TypeAndMut<'tcx>) -> &'tcx TyS<'tcx>

Unstable (rustc_private)

fn mk_ref(&self, r: &'tcx Region, tm: TypeAndMut<'tcx>) -> &'tcx TyS<'tcx>

Unstable (rustc_private)

fn mk_mut_ref(&self, r: &'tcx Region, ty: &'tcx TyS<'tcx>) -> &'tcx TyS<'tcx>

Unstable (rustc_private)

fn mk_imm_ref(&self, r: &'tcx Region, ty: &'tcx TyS<'tcx>) -> &'tcx TyS<'tcx>

Unstable (rustc_private)

fn mk_mut_ptr(&self, ty: &'tcx TyS<'tcx>) -> &'tcx TyS<'tcx>

Unstable (rustc_private)

fn mk_imm_ptr(&self, ty: &'tcx TyS<'tcx>) -> &'tcx TyS<'tcx>

Unstable (rustc_private)

fn mk_nil_ptr(&self) -> &'tcx TyS<'tcx>

Unstable (rustc_private)

fn mk_array(&self, ty: &'tcx TyS<'tcx>, n: usize) -> &'tcx TyS<'tcx>

Unstable (rustc_private)

fn mk_slice(&self, ty: &'tcx TyS<'tcx>) -> &'tcx TyS<'tcx>

Unstable (rustc_private)

fn mk_tup(&self, ts: Vec<&'tcx TyS<'tcx>>) -> &'tcx TyS<'tcx>

Unstable (rustc_private)

fn mk_nil(&self) -> &'tcx TyS<'tcx>

Unstable (rustc_private)

fn mk_bool(&self) -> &'tcx TyS<'tcx>

Unstable (rustc_private)

fn mk_fn(&self, opt_def_id: Option<DefId>, fty: &'tcx BareFnTy<'tcx>) -> &'tcx TyS<'tcx>

Unstable (rustc_private)

fn mk_ctor_fn(&self, def_id: DefId, input_tys: &[&'tcx TyS<'tcx>], output: &'tcx TyS<'tcx>) -> &'tcx TyS<'tcx>

Unstable (rustc_private)

fn mk_trait(&self, principal: Binder<TraitRef<'tcx>>, bounds: ExistentialBounds<'tcx>) -> &'tcx TyS<'tcx>

Unstable (rustc_private)

fn mk_projection(&self, trait_ref: TraitRef<'tcx>, item_name: Name) -> &'tcx TyS<'tcx>

Unstable (rustc_private)

fn mk_struct(&self, def: &'tcx AdtDefData<'tcx, 'static>, substs: &'tcx Substs<'tcx>) -> &'tcx TyS<'tcx>

Unstable (rustc_private)

fn mk_closure(&self, closure_id: DefId, substs: &'tcx Substs<'tcx>, tys: Vec<&'tcx TyS<'tcx>>) -> &'tcx TyS<'tcx>

Unstable (rustc_private)

fn mk_closure_from_closure_substs(&self, closure_id: DefId, closure_substs: Box<ClosureSubsts<'tcx>>) -> &'tcx TyS<'tcx>

Unstable (rustc_private)

fn mk_var(&self, v: TyVid) -> &'tcx TyS<'tcx>

Unstable (rustc_private)

fn mk_int_var(&self, v: IntVid) -> &'tcx TyS<'tcx>

Unstable (rustc_private)

fn mk_float_var(&self, v: FloatVid) -> &'tcx TyS<'tcx>

Unstable (rustc_private)

fn mk_infer(&self, it: InferTy) -> &'tcx TyS<'tcx>

Unstable (rustc_private)

fn mk_param(&self, space: ParamSpace, index: u32, name: Name) -> &'tcx TyS<'tcx>

Unstable (rustc_private)

fn mk_self_type(&self) -> &'tcx TyS<'tcx>

Unstable (rustc_private)

fn mk_param_from_def(&self, def: &TypeParameterDef) -> &'tcx TyS<'tcx>

Unstable (rustc_private)

fn trait_items(&self, trait_did: DefId) -> Rc<Vec<ImplOrTraitItem<'tcx>>>

Unstable (rustc_private)

fn lookup_repr_hints(&self, did: DefId) -> Rc<Vec<ReprAttr>>

Unstable (rustc_private)

Obtain the representation annotation for a struct definition.

impl<'tcx> ctxt<'tcx>

fn try_add_builtin_trait(&self, trait_def_id: DefId, builtin_bounds: &mut EnumSet<BuiltinBound>) -> bool

Unstable (rustc_private)

Checks whether trait_ref refers to one of the builtin traits, like Send, and adds the corresponding bound to the set builtin_bounds if so. Returns true if trait_ref is a builtin trait.

impl<'tcx> ctxt<'tcx>

fn node_id_to_type(&self, id: u32) -> &'tcx TyS<'tcx>

Unstable (rustc_private)

fn node_id_to_type_opt(&self, id: u32) -> Option<&'tcx TyS<'tcx>>

Unstable (rustc_private)

fn node_id_item_substs(&self, id: u32) -> ItemSubsts<'tcx>

Unstable (rustc_private)

fn pat_ty(&self, pat: &Pat) -> &'tcx TyS<'tcx>

Unstable (rustc_private)

fn pat_ty_opt(&self, pat: &Pat) -> Option<&'tcx TyS<'tcx>>

Unstable (rustc_private)

fn expr_ty(&self, expr: &Expr) -> &'tcx TyS<'tcx>

Unstable (rustc_private)

fn expr_ty_opt(&self, expr: &Expr) -> Option<&'tcx TyS<'tcx>>

Unstable (rustc_private)

fn expr_ty_adjusted(&self, expr: &Expr) -> &'tcx TyS<'tcx>

Unstable (rustc_private)

Returns the type of expr, considering any AutoAdjustment entry recorded for that expression.

It would almost certainly be better to store the adjusted ty in with the AutoAdjustment, but I opted not to do this because it would require serializing and deserializing the type and, although that's not hard to do, I just hate that code so much I didn't want to touch it unless it was to fix it properly, which seemed a distraction from the thread at hand! -nmatsakis

fn expr_ty_adjusted_opt(&self, expr: &Expr) -> Option<&'tcx TyS<'tcx>>

Unstable (rustc_private)

fn expr_span(&self, id: u32) -> Span

Unstable (rustc_private)

fn local_var_name_str(&self, id: u32) -> InternedString

Unstable (rustc_private)

fn resolve_expr(&self, expr: &Expr) -> Def

Unstable (rustc_private)

fn expr_is_lval(&self, expr: &Expr) -> bool

Unstable (rustc_private)

fn provided_trait_methods(&self, id: DefId) -> Vec<Rc<Method<'tcx>>>

Unstable (rustc_private)

fn associated_consts(&self, id: DefId) -> Vec<Rc<AssociatedConst<'tcx>>>

Unstable (rustc_private)

fn trait_impl_polarity(&self, id: DefId) -> Option<ImplPolarity>

Unstable (rustc_private)

fn custom_coerce_unsized_kind(&self, did: DefId) -> CustomCoerceUnsized

Unstable (rustc_private)

fn impl_or_trait_item(&self, id: DefId) -> ImplOrTraitItem<'tcx>

Unstable (rustc_private)

fn trait_item_def_ids(&self, id: DefId) -> Rc<Vec<ImplOrTraitItemId>>

Unstable (rustc_private)

fn impl_trait_ref(&self, id: DefId) -> Option<TraitRef<'tcx>>

Unstable (rustc_private)

Returns the trait-ref corresponding to a given impl, or None if it is an inherent impl.

fn is_impl(&self, id: DefId) -> bool

Unstable (rustc_private)

Returns whether this DefId refers to an impl

fn trait_ref_to_def_id(&self, tr: &TraitRef) -> DefId

Unstable (rustc_private)

fn item_path_str(&self, id: DefId) -> String

Unstable (rustc_private)

fn def_path(&self, id: DefId) -> Vec<DisambiguatedDefPathData>

Unstable (rustc_private)

fn with_path<T, F>(&self, id: DefId, f: F) -> T where F: FnOnce(Chain<Cloned<Iter<PathElem>>, LinkedPath>) -> T

Unstable (rustc_private)

fn item_name(&self, id: DefId) -> Name

Unstable (rustc_private)

fn register_item_type(&self, did: DefId, ty: TypeScheme<'tcx>)

Unstable (rustc_private)

fn lookup_item_type(&self, did: DefId) -> TypeScheme<'tcx>

Unstable (rustc_private)

fn lookup_trait_def(&self, did: DefId) -> &'tcx TraitDef<'tcx>

Unstable (rustc_private)

Given the did of a trait, returns its canonical trait ref.

fn lookup_adt_def_master(&self, did: DefId) -> &'tcx AdtDefData<'tcx, 'tcx>

Unstable (rustc_private)

Given the did of an ADT, return a master reference to its definition. Unless you are planning on fulfilling the ADT's fields, use lookup_adt_def instead.

fn lookup_adt_def(&self, did: DefId) -> &'tcx AdtDefData<'tcx, 'static>

Unstable (rustc_private)

Given the did of an ADT, return a reference to its definition.

fn lookup_predicates(&self, did: DefId) -> GenericPredicates<'tcx>

Unstable (rustc_private)

Given the did of an item, returns its full set of predicates.

fn lookup_super_predicates(&self, did: DefId) -> GenericPredicates<'tcx>

Unstable (rustc_private)

Given the did of a trait, returns its superpredicates.

fn type_needs_drop_given_env(&self, ty: &'tcx TyS<'tcx>, param_env: &ParameterEnvironment<'a, 'tcx>) -> bool

Unstable (rustc_private)

If type_needs_drop returns true, then ty is definitely non-copy and might have a destructor attached; if it returns false, then ty definitely has no destructor (i.e. no drop glue).

(Note that this implies that if ty has a destructor attached, then type_needs_drop will definitely return true for ty.)

fn get_attrs(&self, did: DefId) -> Cow<'tcx, [Spanned<Attribute_>]>

Unstable (rustc_private)

Get the attributes of a definition.

fn has_attr(&self, did: DefId, attr: &str) -> bool

Unstable (rustc_private)

Determine whether an item is annotated with an attribute

fn lookup_packed(&self, did: DefId) -> bool

Unstable (rustc_private)

Determine whether an item is annotated with #[repr(packed)]

fn lookup_simd(&self, did: DefId) -> bool

Unstable (rustc_private)

Determine whether an item is annotated with #[simd]

fn item_variances(&self, item_id: DefId) -> Rc<ItemVariances>

Unstable (rustc_private)

fn trait_has_default_impl(&self, trait_def_id: DefId) -> bool

Unstable (rustc_private)

fn record_trait_has_default_impl(&self, trait_def_id: DefId)

Unstable (rustc_private)

Records a trait-to-implementation mapping.

fn populate_implementations_for_primitive_if_necessary(&self, primitive_def_id: DefId)

Unstable (rustc_private)

Load primitive inherent implementations if necessary

fn populate_inherent_implementations_for_type_if_necessary(&self, type_id: DefId)

Unstable (rustc_private)

Populates the type context with all the inherent implementations for the given type if necessary.

fn populate_implementations_for_trait_if_necessary(&self, trait_id: DefId)

Unstable (rustc_private)

Populates the type context with all the implementations for the given trait if necessary.

fn closure_kind(&self, def_id: DefId) -> ClosureKind

Unstable (rustc_private)

fn closure_type(&self, def_id: DefId, substs: &ClosureSubsts<'tcx>) -> ClosureTy<'tcx>

Unstable (rustc_private)

fn trait_id_of_impl(&self, def_id: DefId) -> Option<DefId>

Unstable (rustc_private)

Given the def_id of an impl, return the def_id of the trait it implements. If it implements no trait, return None.

fn impl_of_method(&self, def_id: DefId) -> Option<DefId>

Unstable (rustc_private)

If the given def ID describes a method belonging to an impl, return the ID of the impl that the method belongs to. Otherwise, return None.

fn trait_of_item(&self, def_id: DefId) -> Option<DefId>

Unstable (rustc_private)

If the given def ID describes an item belonging to a trait (either a default method or an implementation of a trait method), return the ID of the trait that the method belongs to. Otherwise, return None.

fn trait_item_of_item(&self, def_id: DefId) -> Option<ImplOrTraitItemId>

Unstable (rustc_private)

If the given def ID describes an item belonging to a trait, (either a default method or an implementation of a trait method), return the ID of the method inside trait definition (this means that if the given def ID is already that of the original trait method, then the return value is the same). Otherwise, return None.

fn empty_parameter_environment(&'a self) -> ParameterEnvironment<'a, 'tcx>

Unstable (rustc_private)

Construct a parameter environment suitable for static contexts or other contexts where there are no free type/lifetime parameters in scope.

fn construct_free_substs(&self, generics: &Generics<'tcx>, free_id_outlive: CodeExtent) -> Substs<'tcx>

Unstable (rustc_private)

Constructs and returns a substitution that can be applied to move from the "outer" view of a type or method to the "inner" view. In general, this means converting from bound parameters to free parameters. Since we currently represent bound/free type parameters in the same way, this only has an effect on regions.

fn construct_parameter_environment(&'a self, span: Span, generics: &Generics<'tcx>, generic_predicates: &GenericPredicates<'tcx>, free_id_outlive: CodeExtent) -> ParameterEnvironment<'a, 'tcx>

Unstable (rustc_private)

See ParameterEnvironment struct def'n for details. If you were using free_id: NodeId, you might try self.region_maps.item_extent(free_id) for the free_id_outlive parameter. (But note that that is not always quite right.)

fn is_method_call(&self, expr_id: u32) -> bool

Unstable (rustc_private)

fn is_overloaded_autoderef(&self, expr_id: u32, autoderefs: u32) -> bool

Unstable (rustc_private)

fn upvar_capture(&self, upvar_id: UpvarId) -> Option<UpvarCapture>

Unstable (rustc_private)

fn visit_all_items_in_krate<V, F>(&self, dep_node_fn: F, visitor: &mut V) where F: FnMut(DefId) -> DepNode, V: Visitor<'tcx>

Unstable (rustc_private)

impl<'tcx> ctxt<'tcx>

fn with_freevars<T, F>(&self, fid: u32, f: F) -> T where F: FnOnce(&[Freevar]) -> T

Unstable (rustc_private)

fn make_substs_for_receiver_types(&self, trait_ref: &TraitRef<'tcx>, method: &Method<'tcx>) -> Substs<'tcx>

Unstable (rustc_private)

Substitutes the values for the receiver's type parameters that are found in method, leaving the method's type parameters intact.