Function rustc_typeck::middle::ty_relate::super_relate_tys

pub fn super_relate_tys<'a, 'tcx, R>(relation: &mut R, a: &'tcx TyS<'tcx>, b: &'tcx TyS<'tcx>) -> Result<&'tcx TyS<'tcx>, type_err<'tcx>> where 'tcx: 'a, R: TypeRelation<'a, 'tcx>

Unstable

The main "type relation" routine. Note that this does not handle inference artifacts, so you should filter those out before calling it.