Trait rustc_typeck::middle::ty_relate::TypeRelation

pub trait TypeRelation<'a, 'tcx> {
    fn tcx(&self) -> &'a ctxt<'tcx>;
    fn tag(&self) -> &'static str;
    fn a_is_expected(&self) -> bool;
    fn relate_with_variance<T>(&mut self, variance: Variance, a: &T, b: &T) -> Result<T, TypeError<'tcx>> where T: Relate<'a, 'tcx>;
    fn tys(&mut self, a: &'tcx TyS<'tcx>, b: &'tcx TyS<'tcx>) -> Result<&'tcx TyS<'tcx>, TypeError<'tcx>>;
    fn regions(&mut self, a: Region, b: Region) -> Result<Region, TypeError<'tcx>>;
    fn binders<T>(&mut self, a: &Binder<T>, b: &Binder<T>) -> Result<Binder<T>, TypeError<'tcx>> where T: Relate<'a, 'tcx>;

    fn with_cause<F, R>(&mut self, _cause: Cause, f: F) -> R where F: FnOnce(&mut Self) -> R { ... }
    fn relate<T>(&mut self, a: &T, b: &T) -> Result<T, TypeError<'tcx>> where T: Relate<'a, 'tcx> { ... }
    fn relate_zip<T>(&mut self, a: &[T], b: &[T]) -> Result<Vec<T>, TypeError<'tcx>> where T: Relate<'a, 'tcx> { ... }
}

Unstable (rustc_private)

Required Methods

fn tcx(&self) -> &'a ctxt<'tcx>

Unstable (rustc_private)

fn tag(&self) -> &'static str

Unstable (rustc_private)

Returns a static string we can use for printouts.

fn a_is_expected(&self) -> bool

Unstable (rustc_private)

Returns true if the value a is the "expected" type in the relation. Just affects error messages.

fn relate_with_variance<T>(&mut self, variance: Variance, a: &T, b: &T) -> Result<T, TypeError<'tcx>> where T: Relate<'a, 'tcx>

Unstable (rustc_private)

Switch variance for the purpose of relating a and b.

fn tys(&mut self, a: &'tcx TyS<'tcx>, b: &'tcx TyS<'tcx>) -> Result<&'tcx TyS<'tcx>, TypeError<'tcx>>

Unstable (rustc_private)

fn regions(&mut self, a: Region, b: Region) -> Result<Region, TypeError<'tcx>>

Unstable (rustc_private)

fn binders<T>(&mut self, a: &Binder<T>, b: &Binder<T>) -> Result<Binder<T>, TypeError<'tcx>> where T: Relate<'a, 'tcx>

Unstable (rustc_private)

Provided Methods

fn with_cause<F, R>(&mut self, _cause: Cause, f: F) -> R where F: FnOnce(&mut Self) -> R

Unstable (rustc_private)

fn relate<T>(&mut self, a: &T, b: &T) -> Result<T, TypeError<'tcx>> where T: Relate<'a, 'tcx>

Unstable (rustc_private)

Generic relation routine suitable for most anything.

fn relate_zip<T>(&mut self, a: &[T], b: &[T]) -> Result<Vec<T>, TypeError<'tcx>> where T: Relate<'a, 'tcx>

Unstable (rustc_private)

Relete elements of two slices pairwise.