Struct rustc_typeck::CrateCtxt

pub struct CrateCtxt<'a, 'tcx: 'a> {
    pub trait_map: TraitMap,
    pub all_traits: RefCell<Option<AllTraitsVec>>,
    pub stack: RefCell<Vec<AstConvRequest>>,
    pub tcx: TyCtxt<'a, 'tcx, 'tcx>,
    // some fields omitted
}

Unstable (rustc_private)

Fields

trait_map
Unstable (rustc_private)

A mapping from method call sites to traits that have that method.

all_traits
Unstable (rustc_private)

A vector of every trait accessible in the whole crate (i.e. including those from subcrates). This is used only for error reporting, and so is lazily initialised and generally shouldn't taint the common path (hence the RefCell).

stack
Unstable (rustc_private)

This stack is used to identify cycles in the user's source. Note that these cycles can cross multiple items.

tcx
Unstable (rustc_private)

Methods

impl<'a, 'gcx, 'tcx> CrateCtxt<'a, 'gcx>

fn inherited(&'a self, param_env: Option<ParameterEnvironment<'gcx>>) -> InheritedBuilder<'a, 'gcx, 'tcx>