Function rustc_typeck::check::_match::check_struct_pat_fields

pub fn check_struct_pat_fields<'a, 'tcx>(pcx: &pat_ctxt<'a, 'tcx>, span: Span, fields: &'tcx [Spanned<FieldPat>], variant: VariantDef<'tcx>, substs: &Substs<'tcx>, etc: bool)

Unstable (rustc_private)

path is the AST path item naming the type of this struct. fields is the field patterns of the struct pattern. struct_fields describes the type of each field of the struct. struct_id is the ID of the struct. etc is true if the pattern said '...' and false otherwise.