Function rustc_typeck::check::dropck::check_safety_of_destructor_if_necessary

pub fn check_safety_of_destructor_if_necessary<'a, 'tcx>(rcx: &mut Rcx<'a, 'tcx>, typ: Ty<'tcx>, span: Span, scope: CodeExtent)

Unstable (rustc_private)

check_safety_of_destructor_if_necessary confirms that the type expression typ conforms to the "Drop Check Rule" from the Sound Generic Drop (RFC 769).

The Drop Check Rule is the following:

Let v be some value (either temporary or named) and 'a be some lifetime (scope). If the type of v owns data of type D, where

(1.) D has a lifetime- or type-parametric Drop implementation, and

(2.) the structure of D can reach a reference of type &'a _, and

(3.) either:

(A.) the Drop impl for D instantiates D at 'a directly, i.e. D<'a>, or,

(B.) the Drop impl for D has some type parameter with a trait bound T where T is a trait that has at least one method,

then 'a must strictly outlive the scope of v.

This function is meant to by applied to the type for every expression in the program.