Function rustc_typeck::check::autoderef

pub fn autoderef<'a, 'tcx, T, F>(fcx: &FnCtxt<'a, 'tcx>, sp: Span, base_ty: Ty<'tcx>, opt_expr: Option<&Expr>, unresolved_type_action: UnresolvedTypeAction, lvalue_pref: LvaluePreference, should_stop: F) -> (Ty<'tcx>, usize, Option<T>) where F: FnMut(Ty<'tcx>, usize) -> Option<T>

Unstable (rustc_private)

Executes an autoderef loop for the type t. At each step, invokes should_stop to decide whether to terminate the loop. Returns the final type and number of derefs that it performed.

Note: this method does not modify the adjustments table. The caller is responsible for inserting an AutoAdjustment record into the fcx using one of the suitable methods.