Struct rustc_typeck::check::FnCtxt

pub struct FnCtxt<'a, 'tcx> {
    // some fields omitted
}

Unstable (rustc_private)

Methods

impl<'a, 'tcx> FnCtxt<'a, 'tcx>

fn infcx(&self) -> &InferCtxt<'a, 'tcx>

Unstable (rustc_private)

fn param_env(&self) -> &ParameterEnvironment<'a, 'tcx>

Unstable (rustc_private)

fn sess(&self) -> &Session

Unstable (rustc_private)

fn err_count_since_creation(&self) -> usize

Unstable (rustc_private)

fn tag(&self) -> String

Unstable (rustc_private)

fn local_ty(&self, span: Span, nid: NodeId) -> Ty<'tcx>

Unstable (rustc_private)

fn write_ty(&self, node_id: NodeId, ty: Ty<'tcx>)

Unstable (rustc_private)

fn write_substs(&self, node_id: NodeId, substs: ItemSubsts<'tcx>)

Unstable (rustc_private)

fn write_autoderef_adjustment(&self, node_id: NodeId, derefs: usize)

Unstable (rustc_private)

fn write_adjustment(&self, node_id: NodeId, adj: AutoAdjustment<'tcx>)

Unstable (rustc_private)

fn instantiate_type(&self, did: DefId, path: &Path) -> Ty<'tcx>

Unstable (rustc_private)

Instantiates the type in did with the generics in path and returns it (registering the necessary trait obligations along the way).

Note that this function is only intended to be used with type-paths, not with value-paths.

fn def_struct_variant(&self, def: Def, span: Span) -> Option<(AdtDef<'tcx>, VariantDef<'tcx>)>

Unstable (rustc_private)

Return the dict-like variant corresponding to a given Def.

fn write_nil(&self, node_id: NodeId)

Unstable (rustc_private)

fn write_error(&self, node_id: NodeId)

Unstable (rustc_private)

fn require_type_meets(&self, ty: Ty<'tcx>, span: Span, code: ObligationCauseCode<'tcx>, bound: BuiltinBound)

Unstable (rustc_private)

fn require_type_is_sized(&self, ty: Ty<'tcx>, span: Span, code: ObligationCauseCode<'tcx>)

Unstable (rustc_private)

fn require_expr_have_sized_type(&self, expr: &Expr, code: ObligationCauseCode<'tcx>)

Unstable (rustc_private)

fn type_is_known_to_be_sized(&self, ty: Ty<'tcx>, span: Span) -> bool

Unstable (rustc_private)

fn register_builtin_bound(&self, ty: Ty<'tcx>, builtin_bound: BuiltinBound, cause: ObligationCause<'tcx>)

Unstable (rustc_private)

fn register_predicate(&self, obligation: PredicateObligation<'tcx>)

Unstable (rustc_private)

fn to_ty(&self, ast_t: &Ty) -> Ty<'tcx>

Unstable (rustc_private)

fn expr_ty(&self, ex: &Expr) -> Ty<'tcx>

Unstable (rustc_private)

fn adjust_expr_ty(&self, expr: &Expr, adjustment: Option<&AutoAdjustment<'tcx>>) -> Ty<'tcx>

Unstable (rustc_private)

Apply adjustment to the type of expr

fn node_ty(&self, id: NodeId) -> Ty<'tcx>

Unstable (rustc_private)

fn item_substs(&self) -> Ref<NodeMap<ItemSubsts<'tcx>>>

Unstable (rustc_private)

fn opt_node_ty_substs<F>(&self, id: NodeId, f: F) where F: FnOnce(&ItemSubsts<'tcx>)

Unstable (rustc_private)

fn mk_subty(&self, a_is_expected: bool, origin: TypeOrigin, sub: Ty<'tcx>, sup: Ty<'tcx>) -> Result<(), TypeError<'tcx>>

Unstable (rustc_private)

fn mk_eqty(&self, a_is_expected: bool, origin: TypeOrigin, sub: Ty<'tcx>, sup: Ty<'tcx>) -> Result<(), TypeError<'tcx>>

Unstable (rustc_private)

fn mk_subr(&self, origin: SubregionOrigin<'tcx>, sub: Region, sup: Region)

Unstable (rustc_private)

fn type_error_message<M>(&self, sp: Span, mk_msg: M, actual_ty: Ty<'tcx>, err: Option<&TypeError<'tcx>>) where M: FnOnce(String) -> String

Unstable (rustc_private)

fn report_mismatched_types(&self, sp: Span, e: Ty<'tcx>, a: Ty<'tcx>, err: &TypeError<'tcx>)

Unstable (rustc_private)

fn register_region_obligation(&self, ty: Ty<'tcx>, region: Region, cause: ObligationCause<'tcx>)

Unstable (rustc_private)

Registers an obligation for checking later, during regionck, that the type ty must outlive the region r.

fn register_wf_obligation(&self, ty: Ty<'tcx>, span: Span, code: ObligationCauseCode<'tcx>)

Unstable (rustc_private)

Registers an obligation for checking later, during regionck, that the type ty must outlive the region r.

fn register_old_wf_obligation(&self, ty: Ty<'tcx>, span: Span, code: ObligationCauseCode<'tcx>)

Unstable (rustc_private)

fn add_wf_bounds(&self, substs: &Substs<'tcx>, expr: &Expr)

Unstable (rustc_private)

Registers obligations that all types appearing in substs are well-formed.

fn add_obligations_for_parameters(&self, cause: ObligationCause<'tcx>, predicates: &InstantiatedPredicates<'tcx>)

Unstable (rustc_private)

Given a fully substituted set of bounds (generic_bounds), and the values with which each type/region parameter was instantiated (substs), creates and registers suitable trait/region obligations.

For example, if there is a function:

fn foo<'a,T:'a>(...)

and a reference:

let f = foo;

Then we will create a fresh region variable '$0 and a fresh type variable $1 for 'a and T. This routine will add a region obligation $1:'$0 and register it locally.

fn field_ty(&self, span: Span, field: FieldDef<'tcx>, substs: &Substs<'tcx>) -> Ty<'tcx>

Unstable (rustc_private)