Struct rustc_typeck::middle::ty::TySUnstable

pub struct TyS<'tcx> {
    pub sty: sty<'tcx>,
    pub flags: TypeFlags,
    // some fields omitted
}

Fields

sty

flags

Methods

impl<'tcx> TyS<'tcx>

fn walk(&'tcx self) -> TypeWalker<'tcx>

Iterator that walks self and any types reachable from self, in depth-first order. Note that just walks the types that appear in self, it does not descend into the fields of structs or variants. For example:

int => { int }
Foo<Bar<int>> => { Foo<Bar<int>>, Bar<int>, int }
[int] => { [int], int }

fn walk_children(&'tcx self) -> TypeWalker<'tcx>

Iterator that walks types reachable from self, in depth-first order. Note that this is a shallow walk. For example:

int => { }
Foo<Bar<int>> => { Bar<int>, int }
[int] => { int }

fn as_opt_param_ty(&self) -> Option<ParamTy>

fn is_param(&self, space: ParamSpace, index: u32) -> bool